Split DocumentWindowController.swift into per-concern extension files
The 3,389-line class becomes a 633-line core (stored properties, window
setup, theming, navigation history, file watching, NSWindowDelegate) plus
ten extensions: Toolbar, Sidebar, AlwaysOnTop, Autosave, EditMode,
FormattingBar, EditSession, OpenTargets (Open / Open-in-LLM / Open-With),
Find, and DocumentOpening. FileWatcher moves to its own file.

Code moves only, except that members used across the new file boundaries
lose `private` — extensions in other files cannot see them. All stored
properties stay in the class body (Swift requires it) and most become
internal for the same reason.
