+++ md-preview/Document/DocumentWindowController.swift
@@ -8,33 +8,6 @@
 import Cocoa
 import UniformTypeIdentifiers
 
-extension NSToolbarItem.Identifier {
-    static let openActions = NSToolbarItem.Identifier("OpenActions")
-    static let openWith = NSToolbarItem.Identifier("OpenWith")
-    static let openInLLM = NSToolbarItem.Identifier("OpenInLLM")
-    static let inspector = NSToolbarItem.Identifier("Inspector")
-    static let share = NSToolbarItem.Identifier("Share")
-    static let search = NSToolbarItem.Identifier("Search")
-    static let sidebarMenu = NSToolbarItem.Identifier("SidebarMenu")
-    static let printDocument = NSToolbarItem.Identifier("PrintDocument")
-    static let exportDocument = NSToolbarItem.Identifier("ExportDocument")
-    static let exportPDF = NSToolbarItem.Identifier("ExportPDF")
-    static let copyMarkdown = NSToolbarItem.Identifier("CopyMarkdown")
-    static let zoom = NSToolbarItem.Identifier("Zoom")
-    static let editDocument = NSToolbarItem.Identifier("EditDocument")
-    static let navigation = NSToolbarItem.Identifier("Navigation")
-    static let alwaysOnTop = NSToolbarItem.Identifier("AlwaysOnTop")
-}
-
-private extension Array where Element == NSToolbarItem.Identifier {
-    mutating func insertAfterOpenActions(_ identifier: NSToolbarItem.Identifier) {
-        guard let index = firstIndex(of: .openActions) else {
-            append(identifier)
-            return
-        }
-        insert(identifier, at: index + 1)
-    }
-}
 
 // `NSMenuItemValidation` is load-bearing, not tidiness. `validateMenuItem` is only
 // reachable from AppKit through an @objc entry point, and `NSWindowController` has
@@ -42,80 +15,80 @@ private extension Array where Element == NSToolbarItem.Identifier {
 // is never called: no menu item ever gets its state, and the failure is silent.
 final class DocumentWindowController: NSWindowController, NSWindowDelegate, NSToolbarDelegate, NSSharingServicePickerToolbarItemDelegate, NSSearchFieldDelegate, NSMenuDelegate, NSMenuItemValidation {
 
-    private enum NavigationIntent {
+    enum NavigationIntent {
         case normal
         case back
         case forward
     }
 
-    private enum DiskFileState {
+    enum DiskFileState {
         case unchanged
         case modified(String)
         case missing
         case unreadable
     }
 
-    private enum EditedMarkdownSaveResult {
+    enum EditedMarkdownSaveResult {
         case saved
         case reloaded(String)
         case cancelled
     }
 
-    private enum UnsavedEditResolution {
+    enum UnsavedEditResolution {
         case save
         case discard
         case cancel
     }
 
-    private enum AutoSaveFeedback: Equatable {
+    enum AutoSaveFeedback: Equatable {
         case none
         case failed
     }
 
-    private struct HistoryEntry {
+    struct HistoryEntry {
         let url: URL
         /// Scroll offset when the user navigated away; restored on back/forward.
         let scrollPosition: CGFloat
     }
 
-    private var currentFileURL: URL?
-    private var currentMarkdown: String?
-    private var backHistory: [HistoryEntry] = []
-    private var forwardHistory: [HistoryEntry] = []
-    private weak var navigationItem: NSToolbarItem?
-    private weak var navigationControl: NSSegmentedControl?
+    var currentFileURL: URL?
+    var currentMarkdown: String?
+    var backHistory: [HistoryEntry] = []
+    var forwardHistory: [HistoryEntry] = []
+    weak var navigationItem: NSToolbarItem?
+    weak var navigationControl: NSSegmentedControl?
     private var fileWatcher: FileWatcher?
-    private var isInspectorToggleSelected = false
-    private weak var openActionsItem: NSMenuToolbarItem?
-    private weak var openWithItem: NSMenuToolbarItem?
-    private weak var openInLLMItem: NSMenuToolbarItem?
-    private weak var inspectorItem: NSToolbarItem?
-    private weak var inspectorButton: NSButton?
-    private weak var alwaysOnTopButton: NSButton?
-    private weak var editItem: NSToolbarItem?
-    private weak var editButton: NSButton?
-    private var editorChangeRevision = 0
+    var isInspectorToggleSelected = false
+    weak var openActionsItem: NSMenuToolbarItem?
+    weak var openWithItem: NSMenuToolbarItem?
+    weak var openInLLMItem: NSMenuToolbarItem?
+    weak var inspectorItem: NSToolbarItem?
+    weak var inspectorButton: NSButton?
+    weak var alwaysOnTopButton: NSButton?
+    weak var editItem: NSToolbarItem?
+    weak var editButton: NSButton?
+    var editorChangeRevision = 0
     /// In-memory source shown by preview before the user saves it.
-    private var editorDraftMarkdown: String?
+    var editorDraftMarkdown: String?
     /// Last known on-disk source, retained while preview displays a draft.
-    private var editorBaselineMarkdown: String?
-    private var isEditorCommitInFlight = false
-    private var pendingEditorCommitRequested = false
-    private var pendingCommitShouldExit = false
-    private var pendingCommitCompletions: [(Bool) -> Void] = []
+    var editorBaselineMarkdown: String?
+    var isEditorCommitInFlight = false
+    var pendingEditorCommitRequested = false
+    var pendingCommitShouldExit = false
+    var pendingCommitCompletions: [(Bool) -> Void] = []
     /// When sidebar navigation starts from edit mode, the newly loaded file
     /// should return to edit mode instead of dropping the user into preview.
-    private var pendingEditModeURL: URL?
-    private var autoSaveTimer: Timer?
-    private var autoSaveTimerID: UUID?
-    private var isPerformingAutomaticSave = false
-    private var autoSaveFeedbackResetWork: DispatchWorkItem?
-    private var autoSaveFeedback = AutoSaveFeedback.none {
+    var pendingEditModeURL: URL?
+    var autoSaveTimer: Timer?
+    var autoSaveTimerID: UUID?
+    var isPerformingAutomaticSave = false
+    var autoSaveFeedbackResetWork: DispatchWorkItem?
+    var autoSaveFeedback = AutoSaveFeedback.none {
         didSet { updateWindowSubtitle() }
     }
     /// Drives the native titlebar subtitle while the editor contains changes
     /// that have not yet been written successfully.
-    private var hasUnsavedEditorChanges = false {
+    var hasUnsavedEditorChanges = false {
         didSet {
             guard oldValue != hasUnsavedEditorChanges else { return }
             if hasUnsavedEditorChanges {
@@ -126,33 +99,33 @@ final class DocumentWindowController: NSWindowController, NSWindowDelegate, NSTo
             updateWindowSubtitle()
         }
     }
-    private weak var editAccessory: NSTitlebarAccessoryViewController?
+    weak var editAccessory: NSTitlebarAccessoryViewController?
     /// The native-style line under the formatting bar while the window is
     /// themed — the .hard scroll edge that normally draws it would paint an
     /// opaque system strip over the theme color, so the line is drawn here
     /// and the edge stays frosted.
-    private weak var editAccessoryHairline: NSView?
-    private weak var copyItem: NSToolbarItem?
-    private var copyFeedbackWork: DispatchWorkItem?
-    private weak var searchField: NSSearchField?
-    private weak var sidebarMenu: NSMenu?
-    private weak var sidebarPopUpButton: NSPopUpButton?
-    private var findBar: FindBar?
-    private var findBarAccessory: NSTitlebarAccessoryViewController?
-    private var searchMode: SearchMode = .contains
-    private var pendingFindWork: DispatchWorkItem?
-    private static let findDebounceDelay: TimeInterval = 0.10
-    private let tableUndoManager = UndoManager()
-    private var isTableUndoSaveInFlight = false
+    weak var editAccessoryHairline: NSView?
+    weak var copyItem: NSToolbarItem?
+    var copyFeedbackWork: DispatchWorkItem?
+    weak var searchField: NSSearchField?
+    weak var sidebarMenu: NSMenu?
+    weak var sidebarPopUpButton: NSPopUpButton?
+    var findBar: FindBar?
+    var findBarAccessory: NSTitlebarAccessoryViewController?
+    var searchMode: SearchMode = .contains
+    var pendingFindWork: DispatchWorkItem?
+    static let findDebounceDelay: TimeInterval = 0.10
+    let tableUndoManager = UndoManager()
+    var isTableUndoSaveInFlight = false
 
-    private var documentWindow: NSWindow {
+    var documentWindow: NSWindow {
         guard let window else {
             fatalError("DocumentWindowController accessed before its window was loaded")
         }
         return window
     }
 
-    private var markdownDocument: MarkdownDocument? {
+    var markdownDocument: MarkdownDocument? {
         document as? MarkdownDocument
     }
 
@@ -177,11 +150,11 @@ final class DocumentWindowController: NSWindowController, NSWindowDelegate, NSTo
     /// One-shot override consumed by the next window's setup: "Open in
     /// New Window" needs that window to skip the native tab group during
     /// its first order-front (when AppKit decides tab placement).
-    private static var nextWindowDeclinesTabbing = false
+    static var nextWindowDeclinesTabbing = false
     /// One-shot for explicit tab requests (Open in New Tab, ⌘T, the tab
     /// bar's "+"): the next window joins the frontmost window's tab group
     /// even when the system tabbing preference wouldn't put it there.
-    private static var nextWindowRequestsTab = false
+    static var nextWindowRequestsTab = false
 
     static func markNextWindowAsSeparate() {
         nextWindowDeclinesTabbing = true
@@ -256,7 +229,7 @@ final class DocumentWindowController: NSWindowController, NSWindowDelegate, NSTo
     /// override. Chrome suppression without a matching tint exposes the
     /// stock fills, so every chrome treatment gates on this, not on
     /// "either scheme customized".
-    private var activeSchemeThemed: Bool {
+    var activeSchemeThemed: Bool {
         let isDark = documentWindow.effectiveAppearance
             .bestMatch(from: [.aqua, .darkAqua]) == .darkAqua
         return ThemeColorsSetting.current
@@ -324,7 +297,7 @@ final class DocumentWindowController: NSWindowController, NSWindowDelegate, NSTo
     /// "Open documents in tabs" preference is on, or when the system
     /// "Prefer tabs when opening documents" setting asks for it. Plain
     /// opens (Finder, ⌘O, recents) otherwise get their own window.
-    private func attachToExistingTabGroupIfNeeded() {
+    func attachToExistingTabGroupIfNeeded() {
         guard !documentWindow.isVisible,
               documentWindow.tabbingMode != .disallowed,
               let host = ([NSApp.mainWindow] + NSApp.orderedWindows)
@@ -426,11 +399,11 @@ final class DocumentWindowController: NSWindowController, NSWindowDelegate, NSTo
         }
     }
 
-    private func present(url: URL) {
+    func present(url: URL) {
         present(url: url, intent: .normal)
     }
 
-    private func present(url: URL, intent: NavigationIntent) {
+    func present(url: URL, intent: NavigationIntent) {
         let fragment = url.fragment?.removingPercentEncoding
         let url = Self.fileURLWithoutFragment(url)
         let preserveEditMode = isEditing || pendingEditModeURL != nil
@@ -445,7 +418,7 @@ final class DocumentWindowController: NSWindowController, NSWindowDelegate, NSTo
         present(url: url, preservingEditMode: preserveEditMode, intent: intent, fragment: fragment)
     }
 
-    private func present(url: URL, preservingEditMode: Bool,
+    func present(url: URL, preservingEditMode: Bool,
                          intent: NavigationIntent, fragment: String?) {
         if url.isExistingDirectory {
             pendingEditModeURL = nil
@@ -563,7 +536,7 @@ final class DocumentWindowController: NSWindowController, NSWindowDelegate, NSTo
     /// Update the open URL and propagate it to the title, recent docs,
     /// Open With list, sidebar selection, and inspector — without
     /// re-rendering the WebView, since the markdown content didn't change.
-    private func handleRename(to newURL: URL) {
+    func handleRename(to newURL: URL) {
         guard currentFileURL != nil else { return }
         currentFileURL = newURL
         markdownDocument?.replaceFileURL(newURL)
@@ -606,2621 +579,10 @@ final class DocumentWindowController: NSWindowController, NSWindowDelegate, NSTo
         }
     }
 
-    func toolbarDefaultItemIdentifiers(_ toolbar: NSToolbar) -> [NSToolbarItem.Identifier] {
-        [
-            .flexibleSpace,
-            .sidebarMenu,
-            .sidebarTrackingSeparator,
-            .navigation,
-            .flexibleSpace,
-            .openActions,
-            .space,
-            .zoom,
-            .inspector,
-            .share,
-            .editDocument,
-            .search
-        ]
-    }
-
-    func toolbarAllowedItemIdentifiers(_ toolbar: NSToolbar) -> [NSToolbarItem.Identifier] {
-        var identifiers: [NSToolbarItem.Identifier] = [
-            .sidebarMenu,
-            .sidebarTrackingSeparator,
-            .navigation,
-            .flexibleSpace,
-            .space,
-            .openActions,
-            .openWith,
-            .editDocument,
-            .inspector,
-            .share,
-            .search,
-            .printDocument,
-            .exportPDF,
-            .exportDocument,
-            .copyMarkdown,
-            .zoom,
-            .alwaysOnTop
-        ]
-        if hasLLMTargetsAvailable {
-            identifiers.insertAfterOpenActions(.openInLLM)
-        }
-        return identifiers
-    }
-
-    func toolbar(_ toolbar: NSToolbar,
-                 itemForItemIdentifier itemIdentifier: NSToolbarItem.Identifier,
-                 willBeInsertedIntoToolbar flag: Bool) -> NSToolbarItem? {
-        switch itemIdentifier {
-        case .sidebarMenu: return makeSidebarMenuItem(willBeInsertedIntoToolbar: flag)
-        case .navigation: return makeNavigationItem()
-        case .openActions: return makeOpenActionsItem()
-        case .openWith: return makeOpenWithItem()
-        case .openInLLM:
-            guard hasLLMTargetsAvailable else { return nil }
-            return makeOpenInLLMItem()
-        case .editDocument: return makeEditItem()
-        case .inspector: return makeInspectorItem()
-        case .alwaysOnTop: return makeAlwaysOnTopItem()
-        case .share: return makeShareItem()
-        case .search: return makeSearchItem()
-        case .printDocument: return makePrintItem()
-        case .exportPDF: return makeExportPDFItem()
-        case .exportDocument: return makeExportItem()
-        case .copyMarkdown: return makeCopyItem()
-        case .zoom: return makeZoomItem()
-        default: return nil
-        }
-    }
-
-    private func makeNavigationItem() -> NSToolbarItem {
-        let item = NSToolbarItem(itemIdentifier: .navigation)
-        let back = NSLocalizedString("Back", comment: "Navigation toolbar back button")
-        let forward = NSLocalizedString("Forward", comment: "Navigation toolbar forward button")
-        item.label = NSLocalizedString("Navigation", comment: "Navigation toolbar item label")
-        item.paletteLabel = NSLocalizedString("Back and Forward", comment: "Navigation toolbar palette label")
-        item.isNavigational = true
-        item.autovalidates = false
-
-        let control = NSSegmentedControl(labels: ["", ""],
-                                         trackingMode: .momentary,
-                                         target: self,
-                                         action: #selector(navigateHistory(_:)))
-        control.segmentStyle = .automatic
-        control.setImage(NSImage(systemSymbolName: "chevron.left", accessibilityDescription: back),
-                         forSegment: 0)
-        control.setImage(NSImage(systemSymbolName: "chevron.right", accessibilityDescription: forward),
-                         forSegment: 1)
-        control.setToolTip(back, forSegment: 0)
-        control.setToolTip(forward, forSegment: 1)
-        item.view = control
-        navigationItem = item
-        navigationControl = control
-        updateNavigationControl()
-        return item
-    }
-
-    @objc private func navigateHistory(_ sender: NSSegmentedControl) {
-        switch sender.selectedSegment {
-        case 0:
-            guard let entry = backHistory.last else { return }
-            present(url: entry.url, intent: .back)
-        case 1:
-            guard let entry = forwardHistory.last else { return }
-            present(url: entry.url, intent: .forward)
-        default:
-            break
-        }
-    }
-
-    private func updateNavigationControl() {
-        let hasHistory = !backHistory.isEmpty || !forwardHistory.isEmpty
-        navigationItem?.isHidden = !hasHistory
-        navigationItem?.isEnabled = hasHistory
-        navigationControl?.isEnabled = hasHistory
-        navigationControl?.setEnabled(!backHistory.isEmpty, forSegment: 0)
-        navigationControl?.setEnabled(!forwardHistory.isEmpty, forSegment: 1)
-    }
-
-    private func makeSidebarMenuItem(willBeInsertedIntoToolbar: Bool) -> NSToolbarItem {
-        let item = NSToolbarItem(itemIdentifier: .sidebarMenu)
-        item.label = NSLocalizedString("Sidebar", comment: "Sidebar toolbar item label")
-        item.paletteLabel = NSLocalizedString("Sidebar", comment: "Sidebar toolbar palette label")
-        item.toolTip = NSLocalizedString("Sidebar options", comment: "Sidebar toolbar item tooltip")
-
-        // NSPopUpButton (pull-down) so a single click anywhere on the button
-        // opens the menu and the chevron renders natively. NSMenuToolbarItem
-        // either splits the click (icon vs chevron) or auto-promotes the first
-        // item out of the dropdown — neither matches the Preview-style pulldown.
-        let popup = NSPopUpButton(frame: .zero, pullsDown: true)
-        popup.translatesAutoresizingMaskIntoConstraints = false
-        popup.bezelStyle = .toolbar
-        popup.imagePosition = .imageOnly
-
-        let menu = NSMenu()
-        menu.identifier = NSUserInterfaceItemIdentifier("SidebarMenu")
-        menu.delegate = self
-        menu.autoenablesItems = false
-        rebuildSidebarMenu(menu)
-        popup.menu = menu
-        popup.sizeToFit()
-
-        let container = NSView()
-        container.translatesAutoresizingMaskIntoConstraints = false
-        container.addSubview(popup)
-        NSLayoutConstraint.activate([
-            popup.leadingAnchor.constraint(equalTo: container.leadingAnchor),
-            popup.trailingAnchor.constraint(equalTo: container.trailingAnchor),
-            popup.centerYAnchor.constraint(equalTo: container.centerYAnchor),
-            container.heightAnchor.constraint(equalToConstant: 32)
-        ])
-
-        item.view = container
-        if willBeInsertedIntoToolbar {
-            sidebarMenu = menu
-            sidebarPopUpButton = popup
-            syncSidebarMenuState()
-        }
-        return item
-    }
-
-    func menuNeedsUpdate(_ menu: NSMenu) {
-        guard menu === sidebarMenu else { return }
-        rebuildSidebarMenu(menu)
-        syncSidebarMenuState()
-    }
-
-    private func rebuildSidebarMenu(_ menu: NSMenu) {
-        menu.removeAllItems()
-
-        // Pull-down NSPopUpButton uses the first item as the always-visible
-        // button face (showing only the icon thanks to imagePosition). The
-        // dropdown shows items 2+, so the button face is reserved here.
-        let face = NSMenuItem()
-        face.image = sidebarFaceImage()
-        menu.addItem(face)
-
-        let hide = NSMenuItem(title: NSLocalizedString("Hide Sidebar", comment: "Sidebar dropdown item"),
-                              action: #selector(hideSidebarFromMenu(_:)),
-                              keyEquivalent: "")
-        hide.target = self
-        menu.addItem(hide)
-
-        let outline = NSMenuItem(title: NSLocalizedString("Table of Contents", comment: "Sidebar dropdown item"),
-                                 action: #selector(selectOutlineMode(_:)),
-                                 keyEquivalent: "")
-        outline.target = self
-        menu.addItem(outline)
-
-        let files = NSMenuItem(title: NSLocalizedString("Project Navigator", comment: "Sidebar dropdown item"),
-                               action: #selector(selectFilesMode(_:)),
-                               keyEquivalent: "")
-        files.target = self
-        menu.addItem(files)
-        syncSidebarMenuState(for: menu)
-    }
-
-    private func syncSidebarMenuState() {
-        if let sidebarMenu {
-            syncSidebarMenuState(for: sidebarMenu)
-        }
-    }
-
-    private func syncSidebarMenuState(for menu: NSMenu) {
-        let state = currentSidebarMenuState()
-        menu.items.first { $0.action == #selector(hideSidebarFromMenu(_:)) }?.state = state.sidebarVisible ? .off : .on
-        menu.items.first { $0.action == #selector(selectOutlineMode(_:)) }?.state = (state.sidebarVisible && state.mode == .outline) ? .on : .off
-        menu.items.first { $0.action == #selector(selectFilesMode(_:)) }?.state = (state.sidebarVisible && state.mode == .files) ? .on : .off
-    }
-
-    var sidebarMenuState: (sidebarVisible: Bool, mode: SidebarViewController.Mode) {
-        currentSidebarMenuState()
-    }
-
-    func reloadPreviewForSettingChange() {
-        (documentWindow.contentViewController as? MainSplitViewController)?
-            .reloadPreviewForSettingChange()
-    }
-
-    func applyTextSizeSetting() {
-        (documentWindow.contentViewController as? MainSplitViewController)?
-            .applyTextSizeSetting()
-    }
-
-    private func currentSidebarMenuState() -> (sidebarVisible: Bool, mode: SidebarViewController.Mode) {
-        let split = documentWindow.contentViewController as? MainSplitViewController
-        let sidebarVisible = split?.isSidebarVisible ?? false
-        let mode = split?.sidebarMode ?? .outline
-        return (sidebarVisible, mode)
-    }
-
-    private func sidebarFaceImage() -> NSImage {
-        let image = NSImage(systemSymbolName: "sidebar.leading",
-                            accessibilityDescription: NSLocalizedString("Sidebar", comment: "Sidebar toolbar image")) ?? NSImage()
-        image.isTemplate = true
-        return image
-    }
-
-    @objc func toggleSidebarFromMenu(_ sender: Any?) {
-        (documentWindow.contentViewController as? MainSplitViewController)?.toggleSidebar()
-        syncSidebarMenuState()
-    }
-
-    @objc func hideSidebarFromMenu(_ sender: Any?) {
-        guard let split = documentWindow.contentViewController as? MainSplitViewController,
-              split.isSidebarVisible else { return }
-        split.toggleSidebar()
-        syncSidebarMenuState()
-    }
-
-    @objc func selectOutlineMode(_ sender: Any?) {
-        guard let split = documentWindow.contentViewController as? MainSplitViewController else { return }
-        split.setSidebarMode(.outline)
-        split.showSidebar()
-        syncSidebarMenuState()
-    }
-
-    @objc func selectFilesMode(_ sender: Any?) {
-        guard let split = documentWindow.contentViewController as? MainSplitViewController else { return }
-        split.setSidebarMode(.files)
-        split.showSidebar()
-        syncSidebarMenuState()
-    }
-
-    /// The toolbar button and the menu item both drive the one app-wide
-    /// preference, so the toggle goes through the app delegate rather than
-    /// changing this window: every open window follows, not just this one.
-    @objc func toggleAlwaysOnTop(_ sender: Any?) {
-        (NSApp.delegate as? AppDelegate)?.applyAlwaysOnTopSetting(!isAlwaysOnTop)
-    }
-
-    /// Takes the current setting: called on every open window when it changes,
-    /// and once on a window being created so it opens already floating.
-    func applyAlwaysOnTopSetting() {
-        applyAlwaysOnTopLevel(isFullScreen: documentWindow.styleMask.contains(.fullScreen))
-        alwaysOnTopButton?.state = isAlwaysOnTop ? .on : .off
-    }
-
-    /// The pin is intent, not the level itself: the level is recomputed on both
-    /// full-screen transitions, so the toolbar toggle stays lit across them and
-    /// the window floats again on the way out.
-    private func applyAlwaysOnTopLevel(isFullScreen: Bool) {
-        documentWindow.level = NSWindow.Level(
-            rawValue: AlwaysOnTopPolicy.windowLevel(isPinned: isAlwaysOnTop,
-                                                    isFullScreen: isFullScreen)
-        )
-    }
-
-    private var isAlwaysOnTop: Bool { AlwaysOnTopPolicy.isEnabled }
-
-    /// The pin icon doubles as the state indicator. An `NSMenuItem` draws its
-    /// image and its check mark in the same leading slot, so beside an icon the
-    /// tick is easy to miss and the item reads as neither on nor off. Filling
-    /// the pin makes the state legible at a glance; `state` is still set, so
-    /// VoiceOver and the menu's own semantics stay correct.
-    private static func alwaysOnTopMenuImage(isPinned: Bool) -> NSImage? {
-        let image = NSImage(
-            systemSymbolName: isPinned ? "pin.fill" : "pin",
-            accessibilityDescription: NSLocalizedString("Always on Top",
-                                                        comment: "Always on Top menu item")
-        )
-        image?.isTemplate = true
-        return image
-    }
-
-    func validateMenuItem(_ menuItem: NSMenuItem) -> Bool {
-        if menuItem.action == #selector(saveDocument(_:)) {
-            return isEditing
-        }
-        if menuItem.action == #selector(toggleAlwaysOnTop(_:)) {
-            menuItem.state = isAlwaysOnTop ? .on : .off
-            menuItem.image = Self.alwaysOnTopMenuImage(isPinned: isAlwaysOnTop)
-            return true
-        }
-        syncSidebarMenuState()
-        return true
-    }
-
-    private func makeInspectorItem() -> NSToolbarItem {
-        let item = NSToolbarItem(itemIdentifier: .inspector)
-        item.label = NSLocalizedString("Inspector", comment: "Inspector toolbar item label")
-        item.paletteLabel = NSLocalizedString("Get Info", comment: "Inspector toolbar palette label")
-        item.toolTip = NSLocalizedString("Show the inspector", comment: "Inspector toolbar item tooltip")
-
-        let button = NSButton(image: inspectorImage(),
-                              target: self,
-                              action: #selector(toggleInspectorAction(_:)))
-        button.setButtonType(.pushOnPushOff)
-        button.toolTip = item.toolTip
-        button.translatesAutoresizingMaskIntoConstraints = false
-
-        let container = NSView()
-        container.translatesAutoresizingMaskIntoConstraints = false
-        container.addSubview(button)
-        NSLayoutConstraint.activate([
-            button.leadingAnchor.constraint(equalTo: container.leadingAnchor, constant: 2),
-            button.topAnchor.constraint(equalTo: container.topAnchor),
-            button.trailingAnchor.constraint(equalTo: container.trailingAnchor, constant: -2),
-            button.bottomAnchor.constraint(equalTo: container.bottomAnchor),
-            button.heightAnchor.constraint(equalToConstant: 32),
-            container.widthAnchor.constraint(equalToConstant: 36),
-            container.heightAnchor.constraint(equalToConstant: 32)
-        ])
-
-        item.view = container
-        inspectorButton = button
-        inspectorItem = item
-        refreshInspectorToggleItem()
-        return item
-    }
-
-    private func makeAlwaysOnTopItem() -> NSToolbarItem {
-        let item = NSToolbarItem(itemIdentifier: .alwaysOnTop)
-        let alwaysOnTop = NSLocalizedString("Always on Top", comment: "Always on Top toolbar item label")
-        item.label = alwaysOnTop
-        item.paletteLabel = alwaysOnTop
-        item.toolTip = NSLocalizedString("Keep Markdown Preview windows in front of other apps",
-                                         comment: "Always on Top toolbar item tooltip")
-
-        let button = NSButton(image: alwaysOnTopImage(),
-                              target: self,
-                              action: #selector(toggleAlwaysOnTop(_:)))
-        button.setButtonType(.pushOnPushOff)
-        button.state = isAlwaysOnTop ? .on : .off
-        button.toolTip = item.toolTip
-        button.translatesAutoresizingMaskIntoConstraints = false
-
-        let container = NSView()
-        container.translatesAutoresizingMaskIntoConstraints = false
-        container.addSubview(button)
-        NSLayoutConstraint.activate([
-            button.leadingAnchor.constraint(equalTo: container.leadingAnchor, constant: 2),
-            button.topAnchor.constraint(equalTo: container.topAnchor),
-            button.trailingAnchor.constraint(equalTo: container.trailingAnchor, constant: -2),
-            button.bottomAnchor.constraint(equalTo: container.bottomAnchor),
-            button.heightAnchor.constraint(equalToConstant: 32),
-            container.widthAnchor.constraint(equalToConstant: 36),
-            container.heightAnchor.constraint(equalToConstant: 32)
-        ])
-
-        item.view = container
-        alwaysOnTopButton = button
-        return item
-    }
-
-    private func alwaysOnTopImage() -> NSImage {
-        let image = NSImage(systemSymbolName: "pin",
-                            accessibilityDescription: NSLocalizedString("Always on Top", comment: "Always on Top toolbar image")) ?? NSImage()
-        image.isTemplate = true
-        return image
-    }
-
-    private func makeShareItem() -> NSToolbarItem {
-        let item = NSSharingServicePickerToolbarItem(itemIdentifier: .share)
-        item.label = NSLocalizedString("Share", comment: "Share toolbar item label")
-        item.paletteLabel = NSLocalizedString("Share", comment: "Share toolbar palette label")
-        item.toolTip = NSLocalizedString("Share document", comment: "Share toolbar item tooltip")
-        item.delegate = self
-        return item
-    }
-
-    private func makePrintItem() -> NSToolbarItem {
-        let item = NSToolbarItem(itemIdentifier: .printDocument)
-        let print = NSLocalizedString("Print", comment: "Print toolbar item label")
-        item.label = print
-        item.paletteLabel = print
-        item.toolTip = NSLocalizedString("Print document", comment: "Print toolbar item tooltip")
-        item.image = NSImage(systemSymbolName: "printer",
-                             accessibilityDescription: print)
-        item.isBordered = true
-        item.action = #selector(MainSplitViewController.printMarkdown(_:))
-        return item
-    }
-
-    private func makeExportPDFItem() -> NSToolbarItem {
-        let item = NSToolbarItem(itemIdentifier: .exportPDF)
-        let label = NSLocalizedString(
-            "Export PDF", comment: "Export as PDF toolbar item label")
-        item.label = label
-        item.paletteLabel = label
-        item.toolTip = NSLocalizedString(
-            "Export document as PDF", comment: "Export as PDF toolbar item tooltip")
-        item.image = NSImage(systemSymbolName: "arrow.up.document",
-                             accessibilityDescription: label)
-        item.isBordered = true
-        item.action = #selector(MainSplitViewController.exportMarkdownAsPDF(_:))
-        return item
-    }
-
-    private func makeExportItem() -> NSToolbarItem {
-        let item = NSToolbarItem(itemIdentifier: .exportDocument)
-        let label = NSLocalizedString(
-            "Export", comment: "Export toolbar item label")
-        item.label = label
-        item.paletteLabel = label
-        item.toolTip = NSLocalizedString(
-            "Export document", comment: "Export toolbar item tooltip")
-        item.image = NSImage(systemSymbolName: "document.badge.arrow.up",
-                             accessibilityDescription: label)
-        item.isBordered = true
-        item.action = #selector(MainSplitViewController.exportMarkdownDocument(_:))
-        return item
-    }
-
-    private func makeCopyItem() -> NSToolbarItem {
-        let item = NSToolbarItem(itemIdentifier: .copyMarkdown)
-        let copy = NSLocalizedString("Copy", comment: "Copy toolbar item label")
-        item.label = copy
-        item.paletteLabel = copy
-        item.toolTip = NSLocalizedString("Copy Markdown source to clipboard", comment: "Copy toolbar item tooltip")
-        item.image = copyIdleImage()
-        item.isBordered = true
-        item.target = self
-        item.action = #selector(copyMarkdownAction(_:))
-        copyItem = item
-        return item
-    }
-
-    @objc private func copyMarkdownAction(_ sender: Any?) {
-        guard let markdown = currentMarkdown, !markdown.isEmpty else {
-            NSSound.beep()
-            return
-        }
-        let pasteboard = NSPasteboard.general
-        pasteboard.clearContents()
-        pasteboard.setString(markdown, forType: .string)
-        flashCopyFeedback()
-    }
-
-    private static let copyFeedbackDuration: TimeInterval = 1.2
-
-    private func flashCopyFeedback() {
-        guard let item = copyItem else { return }
-        copyFeedbackWork?.cancel()
-        item.image = copyConfirmedImage()
-        let work = DispatchWorkItem { [weak self] in
-            self?.copyItem?.image = self?.copyIdleImage()
-        }
-        copyFeedbackWork = work
-        DispatchQueue.main.asyncAfter(
-            deadline: .now() + Self.copyFeedbackDuration, execute: work
-        )
-    }
-
-    // MARK: - Autosave
-
-    func applyAutoSaveIntervalSetting() {
-        guard hasUnsavedEditorChanges else { return }
-        stopAutoSaveTimer()
-        startAutoSaveTimerIfNeeded()
-    }
-
-    private func startAutoSaveTimerIfNeeded() {
-        guard autoSaveTimer == nil,
-              currentFileURL != nil,
-              hasUnsavedEditorChanges,
-              let interval = AutoSaveSetting.interval else { return }
-        let timerID = UUID()
-        autoSaveTimerID = timerID
-        autoSaveTimer = Timer.scheduledTimer(
-            withTimeInterval: interval,
-            repeats: false
-        ) { [weak self, timerID] _ in
-            Task { @MainActor [weak self] in
-                guard let self, self.autoSaveTimerID == timerID else { return }
-                self.autoSaveTimer = nil
-                self.autoSaveTimerID = nil
-                self.performAutomaticSave()
-            }
-        }
-    }
-
-    private func stopAutoSaveTimer() {
-        autoSaveTimer?.invalidate()
-        autoSaveTimer = nil
-        autoSaveTimerID = nil
-    }
-
-    private func performAutomaticSave() {
-        guard hasUnsavedEditorChanges,
-              !isEditorCommitInFlight,
-              currentFileURL != nil else { return }
-
-        isPerformingAutomaticSave = true
-        commitEdits(exitAfter: false) { [weak self] success in
-            guard let self else { return }
-            self.isPerformingAutomaticSave = false
-            guard !success else { return }
-            self.showAutoSaveFailure()
-        }
-    }
-
-    private func showAutoSaveFailure() {
-        autoSaveFeedbackResetWork?.cancel()
-        autoSaveFeedbackResetWork = nil
-        autoSaveFeedback = .failed
-
-        let work = DispatchWorkItem { [weak self] in
-            guard let self else { return }
-            self.autoSaveFeedbackResetWork = nil
-            self.autoSaveFeedback = .none
-        }
-        autoSaveFeedbackResetWork = work
-        DispatchQueue.main.asyncAfter(deadline: .now() + 1.5, execute: work)
-    }
-
-    private func resetAutoSaveFeedback() {
-        autoSaveFeedbackResetWork?.cancel()
-        autoSaveFeedbackResetWork = nil
-        autoSaveFeedback = .none
-    }
-
-    private func updateWindowSubtitle() {
-        switch autoSaveFeedback {
-        case .failed:
-            documentWindow.subtitle = NSLocalizedString(
-                "Auto-save failed", comment: "Window subtitle after an automatic save failure")
-        case .none:
-            documentWindow.subtitle = hasUnsavedEditorChanges
-                ? NSLocalizedString("Edited", comment: "Window subtitle for unsaved changes")
-                : ""
-        }
-    }
-
-    private func copyIdleImage() -> NSImage? {
-        NSImage(systemSymbolName: "document.on.document",
-                accessibilityDescription: NSLocalizedString("Copy", comment: "Copy toolbar image"))
-    }
-
-    private func copyConfirmedImage() -> NSImage? {
-        NSImage(systemSymbolName: "checkmark",
-                accessibilityDescription: NSLocalizedString("Copied", comment: "Copy confirmation image"))
-    }
-
-    // MARK: - Edit mode
-
-    private var mainSplit: MainSplitViewController? {
-        documentWindow.contentViewController as? MainSplitViewController
-    }
-
-    private var isEditing: Bool {
-        mainSplit?.isEditingDocument ?? false
-    }
-
-    var canToggleEditMode: Bool {
-        isEditing || (currentFileURL != nil && currentMarkdown != nil)
-    }
-
-    var canFormatMarkdown: Bool { isEditing }
-
-    func formatMarkdown(_ command: String) {
-        guard isEditing else { return }
-        mainSplit?.editorViewController?.exec(command)
-    }
-
-    var hasPendingEditorChanges: Bool {
-        hasUnsavedEditorChanges || isEditorCommitInFlight
-    }
-
-    func commitPendingEditsForTermination(completion: @escaping (Bool) -> Void) {
-        guard isEditing || hasPendingEditorChanges else {
-            completion(true)
-            return
-        }
-        requestEndEditing(completion: completion)
-    }
-
-    private func makeEditItem() -> NSToolbarItem {
-        let item = NSToolbarItem(itemIdentifier: .editDocument)
-        let edit = NSLocalizedString("Edit", comment: "Edit toolbar item label")
-        item.label = edit
-        item.paletteLabel = edit
-        item.toolTip = NSLocalizedString("Edit document", comment: "Edit toolbar item tooltip")
-
-        let image = NSImage(systemSymbolName: "highlighter",
-                            accessibilityDescription: edit) ?? NSImage()
-        image.isTemplate = true
-        let button = NSButton(image: image,
-                              target: self,
-                              action: #selector(toggleEditAction(_:)))
-        button.setButtonType(.pushOnPushOff)
-        button.toolTip = item.toolTip
-        button.translatesAutoresizingMaskIntoConstraints = false
-
-        let container = NSView()
-        container.translatesAutoresizingMaskIntoConstraints = false
-        container.addSubview(button)
-        NSLayoutConstraint.activate([
-            button.leadingAnchor.constraint(equalTo: container.leadingAnchor, constant: 2),
-            button.topAnchor.constraint(equalTo: container.topAnchor),
-            button.trailingAnchor.constraint(equalTo: container.trailingAnchor, constant: -2),
-            button.bottomAnchor.constraint(equalTo: container.bottomAnchor),
-            button.heightAnchor.constraint(equalToConstant: 32),
-            container.widthAnchor.constraint(equalToConstant: 36),
-            container.heightAnchor.constraint(equalToConstant: 32)
-        ])
-
-        item.view = container
-        editButton = button
-        editItem = item
-        updateEditToolbarItem()
-        return item
-    }
-
-    private func updateEditToolbarItem() {
-        let editing = isEditing
-        editButton?.state = editing ? .on : .off
-        editItem?.toolTip = editing
-            ? NSLocalizedString("Stop editing and return to preview", comment: "Edit toolbar item tooltip while editing")
-            : NSLocalizedString("Edit document", comment: "Edit toolbar item tooltip")
-        editButton?.toolTip = editItem?.toolTip
-        editButton?.isEnabled = editing || (currentFileURL != nil && currentMarkdown != nil)
-    }
-
-    @objc private func toggleEditAction(_ sender: Any?) {
-        toggleEditMode()
-    }
-
-    func toggleEditMode() {
-        if isEditing {
-            previewPendingEdits()
-        } else {
-            enterEditMode()
-        }
-    }
-
-    // MARK: Formatting bar
-
-    /// Second toolbar row with common markdown actions, shown while
-    /// editing — like Preview's markup bar.
-    private func showEditAccessory() {
-        guard editAccessory == nil else { return }
-
-        let symbolConfig = NSImage.SymbolConfiguration(pointSize: 12, weight: .medium)
-        func formatButton(_ symbol: String, _ command: String, _ tip: String) -> NSButton {
-            let image = NSImage(systemSymbolName: symbol, accessibilityDescription: tip)?
-                .withSymbolConfiguration(symbolConfig) ?? NSImage()
-            let button = NSButton(image: image, target: self, action: #selector(formatCommand(_:)))
-            button.identifier = NSUserInterfaceItemIdentifier(command)
-            // Preview-style: small bare icons, bezel only under the pointer.
-            button.bezelStyle = .accessoryBarAction
-            button.controlSize = .small
-            button.showsBorderOnlyWhileMouseInside = true
-            button.toolTip = tip
-            // The accessory-bar bezel pads the icon generously; a fixed
-            // width tightens the leading/trailing space around the glyph.
-            button.translatesAutoresizingMaskIntoConstraints = false
-            button.widthAnchor.constraint(equalToConstant: 26).isActive = true
-            return button
-        }
-
-        // Plain button with a composed icon+chevron face: unlike a pull-down,
-        // the chevron stays visible when the hover-only bezel is hidden.
-        let headingTitle = NSLocalizedString("Heading", comment: "Formatting toolbar heading button")
-        let headingIcon = NSImage(systemSymbolName: "textformat.size",
-                                  accessibilityDescription: headingTitle)?
-            .withSymbolConfiguration(symbolConfig) ?? NSImage()
-        let headingChevron = NSImage(systemSymbolName: "chevron.down",
-                                     accessibilityDescription: nil)?
-            .withSymbolConfiguration(NSImage.SymbolConfiguration(pointSize: 8, weight: .semibold)) ?? NSImage()
-        let gap: CGFloat = 4
-        let faceSize = NSSize(width: headingIcon.size.width + gap + headingChevron.size.width,
-                              height: max(headingIcon.size.height, headingChevron.size.height))
-        let headingFace = NSImage(size: faceSize, flipped: false) { _ in
-            headingIcon.draw(at: NSPoint(x: 0, y: (faceSize.height - headingIcon.size.height) / 2),
-                             from: .zero, operation: .sourceOver, fraction: 1)
-            headingChevron.draw(at: NSPoint(x: headingIcon.size.width + gap,
-                                            y: (faceSize.height - headingChevron.size.height) / 2),
-                                from: .zero, operation: .sourceOver, fraction: 1)
-            return true
-        }
-        headingFace.isTemplate = true
-        let headings = NSButton(image: headingFace, target: self,
-                                action: #selector(showHeadingMenu(_:)))
-        headings.bezelStyle = .accessoryBarAction
-        headings.controlSize = .small
-        headings.showsBorderOnlyWhileMouseInside = true
-        headings.toolTip = headingTitle
-        headings.translatesAutoresizingMaskIntoConstraints = false
-        headings.widthAnchor.constraint(equalToConstant: 44).isActive = true
-
-        let views: [NSView] = [
-            headings,
-            separatorView(),
-            formatButton("bold", "bold", NSLocalizedString("Bold", comment: "Formatting toolbar tooltip")),
-            formatButton("italic", "italic", NSLocalizedString("Italic", comment: "Formatting toolbar tooltip")),
-            formatButton("strikethrough", "strikethrough", NSLocalizedString("Strikethrough", comment: "Formatting toolbar tooltip")),
-            separatorView(),
-            formatButton("list.bullet", "bulletList", NSLocalizedString("Bulleted List", comment: "Formatting toolbar tooltip")),
-            formatButton("list.number", "orderedList", NSLocalizedString("Numbered List", comment: "Formatting toolbar tooltip")),
-            formatButton("checklist", "taskList", NSLocalizedString("Task List", comment: "Formatting toolbar tooltip")),
-            formatButton("text.quote", "quote", NSLocalizedString("Block Quote", comment: "Formatting toolbar tooltip")),
-            separatorView(),
-            formatButton("chevron.left.forwardslash.chevron.right", "code", NSLocalizedString("Inline Code", comment: "Formatting toolbar tooltip")),
-            formatButton("link", "link", NSLocalizedString("Link", comment: "Formatting toolbar tooltip")),
-        ]
-        let stack = NSStackView(views: views)
-        stack.orientation = .horizontal
-        stack.spacing = 2
-        // Buttons stay tight (2px); the group dividers get room to breathe.
-        for (index, view) in views.enumerated() where view is NSBox {
-            if index > 0 { stack.setCustomSpacing(8, after: views[index - 1]) }
-            stack.setCustomSpacing(8, after: view)
-        }
-        stack.edgeInsets = NSEdgeInsets(top: 4, left: 12, bottom: 6, right: 12)
-        stack.translatesAutoresizingMaskIntoConstraints = false
-
-        // Translucent like the main toolbar: nothing can render behind the
-        // bar (the editor scroller clips text at its hairline and the pocket
-        // ends at the toolbar), so the bar needs no opaque backing — the
-        // flat theme color shows through. Without a theme the .hard edge
-        // still paints the classic opaque bar.
-        let container = EditAccessoryContainerView()
-        container.addSubview(stack)
-        NSLayoutConstraint.activate([
-            stack.leadingAnchor.constraint(equalTo: container.leadingAnchor),
-            stack.trailingAnchor.constraint(lessThanOrEqualTo: container.trailingAnchor),
-            stack.topAnchor.constraint(equalTo: container.topAnchor),
-            stack.bottomAnchor.constraint(equalTo: container.bottomAnchor),
-        ])
-        let hairline = NSBox()
-        hairline.boxType = .separator
-        hairline.translatesAutoresizingMaskIntoConstraints = false
-        container.addSubview(hairline)
-        NSLayoutConstraint.activate([
-            hairline.leadingAnchor.constraint(equalTo: container.leadingAnchor),
-            hairline.trailingAnchor.constraint(equalTo: container.trailingAnchor),
-            hairline.bottomAnchor.constraint(equalTo: container.bottomAnchor),
-        ])
-        hairline.isHidden = !activeSchemeThemed
-        editAccessoryHairline = hairline
-
-        let accessory = NSTitlebarAccessoryViewController()
-        accessory.view = container
-        accessory.layoutAttribute = .bottom
-        accessory.fullScreenMinHeight = 34
-        // macOS 26 replaced the titlebar separator with scroll edge
-        // effects; hard = the classic line under the bar. With a themed
-        // window background the hard edge would paint an opaque system
-        // strip over the theme color, so the frosted automatic style is
-        // used instead.
-        if #available(macOS 26.1, *) {
-            accessory.preferredScrollEdgeEffectStyle =
-                activeSchemeThemed ? .automatic : .hard
-        }
-        documentWindow.addTitlebarAccessoryViewController(accessory)
-        editAccessory = accessory
-    }
-
-    private func separatorView() -> NSView {
-        let line = NSBox()
-        line.boxType = .separator
-        line.translatesAutoresizingMaskIntoConstraints = false
-        line.heightAnchor.constraint(equalToConstant: 16).isActive = true
-        return line
-    }
-
-    private func hideEditAccessory() {
-        guard let accessory = editAccessory else { return }
-        accessory.removeFromParent()
-        editAccessory = nil
-    }
-
-    /// Leaves edit-mode chrome as one operation: drops the formatting bar
-    /// and refreshes the toolbar pencil state.
-    private func dismissEditChrome() {
-        hideEditAccessory()
-        updateEditToolbarItem()
-    }
-
-    @objc private func formatCommand(_ sender: NSButton) {
-        guard let command = sender.identifier?.rawValue else { return }
-        formatMarkdown(command)
-    }
-
-    @objc private func showHeadingMenu(_ sender: NSButton) {
-        let menu = NSMenu()
-        let normalText = NSMenuItem(title: NSLocalizedString("Normal Text", comment: "Formatting toolbar heading menu"),
-                                    action: #selector(headingCommand(_:)),
-                                    keyEquivalent: "")
-        normalText.target = self
-        normalText.tag = 0
-        menu.addItem(normalText)
-        menu.addItem(.separator())
-        for level in 1...3 {
-            let item = NSMenuItem(title: String(
-                format: NSLocalizedString("Heading %d", comment: "Formatting toolbar heading menu level"),
-                level
-            ),
-                                  action: #selector(headingCommand(_:)),
-                                  keyEquivalent: "")
-            item.target = self
-            item.tag = level
-            menu.addItem(item)
-        }
-        menu.popUp(positioning: nil,
-                   at: NSPoint(x: 0, y: sender.bounds.maxY + 4),
-                   in: sender)
-    }
-
-    @objc private func headingCommand(_ sender: NSMenuItem) {
-        formatMarkdown("h\(sender.tag)")
-    }
-
-    /// File > Save (⌘S) while editing: write without leaving edit mode.
-    /// Intercepts the responder chain ahead of MarkdownDocument, whose
-    /// NSDocument save machinery stays disabled.
-    @IBAction func saveDocument(_ sender: Any?) {
-        guard isEditing || hasPendingEditorChanges else {
-            NSSound.beep()
-            return
-        }
-        commitEdits(exitAfter: false)
-    }
-
-    private func enterEditMode() {
-        guard let split = mainSplit, !split.isEditingDocument,
-              currentFileURL != nil,
-              let markdown = editorDraftMarkdown ?? currentMarkdown else {
-            NSSound.beep()
-            return
-        }
-
-        // Edit the complete source. Frontmatter is stripped only by the
-        // read-only renderer; the editor must expose and preserve it.
-        let editor = split.enterEditMode(markdown: markdown)
-        editor.cancelRequested = { [weak self] in
-            self?.previewPendingEdits()
-        }
-        editor.contentDidChange = { [weak self] in
-            self?.editorChangeRevision += 1
-            self?.hasUnsavedEditorChanges = true
-            self?.stopAutoSaveTimer()
-            self?.startAutoSaveTimerIfNeeded()
-        }
-        if editorBaselineMarkdown == nil {
-            editorBaselineMarkdown = currentMarkdown
-            editorChangeRevision = 0
-            hasUnsavedEditorChanges = false
-        }
-        showEditAccessory()
-        updateEditToolbarItem()
-    }
-
-    /// Switches to preview without resolving the editing session. The preview
-    /// renders the in-memory source, while the disk baseline is kept for a
-    /// later Save or close decision.
-    private func previewPendingEdits() {
-        guard let split = mainSplit, let editor = split.editorViewController else { return }
-        editor.fetchMarkdown { [weak self] markdown in
-            guard let self, let markdown, let url = self.currentFileURL else {
-                NSSound.beep()
-                return
-            }
-            self.editorDraftMarkdown = markdown
-            self.currentMarkdown = markdown
-            self.hasUnsavedEditorChanges = markdown != self.editorBaselineMarkdown
-            if !self.hasUnsavedEditorChanges {
-                self.editorDraftMarkdown = nil
-                self.editorBaselineMarkdown = nil
-            }
-            self.markdownDocument?.replaceContents(markdown: markdown, fileURL: url)
-            // exitEditMode(rerender: true) renders the pending markdown once
-            // the editor's scroll anchor has been captured; rendering here as
-            // well raced the anchor hand-off and re-laid the preview out
-            // twice, which showed as jitter during the mode switch. The
-            // formatting accessory likewise stays mounted until the overlay
-            // has faded — removing it earlier reflows the content area in
-            // the middle of the crossfade.
-            self.exitEditMode(rerender: true,
-                              preserveUnsavedChanges: true,
-                              hidesAccessoryAfterFade: true) {}
-        }
-    }
-
-    private func requestEndEditing(
-        keepAccessoryMounted: Bool = false,
-        completion: ((Bool) -> Void)? = nil
-    ) {
-        let finish: (Bool) -> Void = { [weak self] success in
-            if success, !keepAccessoryMounted {
-                self?.dismissEditChrome()
-            }
-            completion?(success)
-        }
-        resolveUnsavedEdits { [weak self] resolution in
-            guard let self else {
-                finish(false)
-                return
-            }
-            switch resolution {
-            case .save:
-                self.commitEdits(exitAfter: true, completion: finish)
-            case .discard:
-                self.exitEditModeWithoutSaving {
-                    finish(true)
-                }
-            case .cancel:
-                finish(false)
-            }
-        }
-    }
-
-    private func resolveUnsavedEdits(
-        completion: @escaping (UnsavedEditResolution) -> Void
-    ) {
-        guard hasUnsavedEditorChanges || isEditorCommitInFlight else {
-            completion(.discard)
-            return
-        }
-
-        // If an explicit ⌘S is already running, let that request finish
-        // instead of presenting a second decision on top of it.
-        if isEditorCommitInFlight {
-            commitEdits(exitAfter: false) { success in
-                completion(success ? .discard : .cancel)
-            }
-            return
-        }
-
-        let alert = NSAlert()
-        alert.alertStyle = .warning
-        alert.messageText = NSLocalizedString(
-            "Do you want to save your changes?",
-            comment: "Unsaved changes alert title"
-        )
-        let fileName = currentFileURL?.lastPathComponent
-            ?? NSLocalizedString("this document", comment: "Fallback document name in alerts")
-        alert.informativeText = String(
-            format: NSLocalizedString(
-                "Your changes to %@ will be lost if you don’t save them.",
-                comment: "Unsaved changes alert message"
-            ),
-            fileName
-        )
-        alert.addButton(withTitle: NSLocalizedString("Save", comment: "Alert button"))
-        alert.addButton(withTitle: NSLocalizedString("Cancel", comment: "Alert button"))
-        alert.addButton(withTitle: NSLocalizedString("Don’t Save", comment: "Alert button"))
-        alert.beginSheetModal(for: documentWindow) { response in
-            switch response {
-            case .alertFirstButtonReturn:
-                completion(.save)
-            case .alertThirdButtonReturn:
-                completion(.discard)
-            default:
-                completion(.cancel)
-            }
-        }
-    }
-
-    private func exitEditModeWithoutSaving(completion: @escaping () -> Void) {
-        var shouldRerender = false
-        if let baseline = editorBaselineMarkdown {
-            currentMarkdown = baseline
-            if let url = currentFileURL {
-                markdownDocument?.replaceContents(markdown: baseline, fileURL: url)
-            }
-            shouldRerender = true
-        }
-        if case let .modified(externalMarkdown) = diskFileState(
-            for: currentFileURL,
-            expectedMarkdown: editorBaselineMarkdown ?? currentMarkdown
-        ) {
-            currentMarkdown = externalMarkdown
-            if let url = currentFileURL {
-                markdownDocument?.replaceContents(markdown: externalMarkdown, fileURL: url)
-            }
-            shouldRerender = true
-        }
-        editorDraftMarkdown = nil
-        editorBaselineMarkdown = nil
-        hasUnsavedEditorChanges = false
-        exitEditMode(rerender: shouldRerender, completion: completion)
-    }
-
-    /// Serializes the editor, writes the file if the content changed, and
-    /// optionally returns to the preview. Stays in edit mode when the save
-    /// fails so no edits are lost. `completion(false)` means the commit
-    /// did not go through.
-    private func commitEdits(exitAfter: Bool, completion: ((Bool) -> Void)? = nil) {
-        pendingCommitShouldExit = pendingCommitShouldExit || exitAfter
-        if let completion {
-            pendingCommitCompletions.append(completion)
-        }
-        guard !isEditorCommitInFlight else {
-            pendingEditorCommitRequested = true
-            return
-        }
-        performPendingEditorCommit()
-    }
-
-    private func performPendingEditorCommit() {
-        let exitAfter = pendingCommitShouldExit
-        pendingCommitShouldExit = false
-        guard let split = mainSplit else {
-            finishEditorCommit(success: false)
-            return
-        }
-        if !isEditing, let body = editorDraftMarkdown {
-            performPendingEditorCommit(body: body, editor: nil, exitAfter: exitAfter)
-            return
-        }
-        guard let editor = split.editorViewController else {
-            finishEditorCommit(success: true)
-            return
-        }
-        let revision = editorChangeRevision
-        isEditorCommitInFlight = true
-        editor.fetchMarkdown { [weak self] body in
-            guard let self else {
-                return
-            }
-            guard let body else {
-                NSSound.beep()
-                self.finishEditorCommit(success: false)
-                return
-            }
-            self.performPendingEditorCommit(body: body, editor: editor,
-                                            revision: revision, exitAfter: exitAfter)
-        }
-    }
-
-    private func performPendingEditorCommit(body: String,
-                                            editor: EditorViewController?,
-                                            revision: Int? = nil,
-                                            exitAfter: Bool) {
-        isEditorCommitInFlight = true
-        let baseline = editorBaselineMarkdown ?? currentMarkdown
-        let diskState = diskFileState(for: currentFileURL,
-                                      expectedMarkdown: baseline)
-        let hasLocalChanges = body != baseline
-
-        if !hasLocalChanges, case let .modified(externalMarkdown) = diskState {
-            // Nothing local needs preserving, so adopt the newer disk
-            // version without presenting a needless conflict dialog.
-            adoptExternalMarkdown(externalMarkdown,
-                                  editor: editor,
-                                  exitAfter: exitAfter)
-            return
-        }
-
-        guard hasUnsavedEditorChanges, hasLocalChanges else {
-            if revision == nil || revision == editorChangeRevision {
-                hasUnsavedEditorChanges = false
-            }
-            switch diskState {
-            case .unchanged:
-                completeSuccessfulEditorCommit(exitAfter: exitAfter,
-                                               rerender: false)
-            case .modified:
-                // Handled above.
-                break
-            case .missing, .unreadable:
-                saveEditedMarkdown(body, diskState: diskState) { result in
-                    self.handleEditedMarkdownSaveResult(result,
-                                                        body: body,
-                                                        editor: editor,
-                                                        revision: revision,
-                                                        exitAfter: exitAfter)
-                }
-            }
-            return
-        }
-        saveEditedMarkdown(body, diskState: diskState) { result in
-            self.handleEditedMarkdownSaveResult(result,
-                                                body: body,
-                                                editor: editor,
-                                                revision: revision,
-                                                exitAfter: exitAfter)
-        }
-    }
-
-    private func handleEditedMarkdownSaveResult(_ result: EditedMarkdownSaveResult,
-                                                body: String,
-                                                editor: EditorViewController?,
-                                                revision: Int?,
-                                                exitAfter: Bool) {
-        switch result {
-        case .saved:
-            currentMarkdown = body
-            if let url = currentFileURL {
-                markdownDocument?.replaceContents(markdown: body, fileURL: url)
-            }
-            editorDraftMarkdown = nil
-            editorBaselineMarkdown = isEditing && !exitAfter ? body : nil
-            if revision == nil || revision == editorChangeRevision {
-                hasUnsavedEditorChanges = false
-            }
-            completeSuccessfulEditorCommit(exitAfter: exitAfter, rerender: true)
-        case let .reloaded(externalMarkdown):
-            adoptExternalMarkdown(externalMarkdown,
-                                  editor: editor,
-                                  exitAfter: exitAfter)
-        case .cancelled:
-            finishEditorCommit(success: false)
-        }
-    }
-
-    private func adoptExternalMarkdown(_ markdown: String,
-                                       editor: EditorViewController?,
-                                       exitAfter: Bool) {
-        currentMarkdown = markdown
-        editorDraftMarkdown = nil
-        editorBaselineMarkdown = isEditing && !exitAfter ? markdown : nil
-        editorChangeRevision = 0
-        hasUnsavedEditorChanges = false
-        if let url = currentFileURL {
-            markdownDocument?.replaceContents(markdown: markdown, fileURL: url)
-            if !exitAfter {
-                renderCurrentDocument(text: markdown, fileURL: url)
-            }
-        }
-        if !exitAfter {
-            editor?.load(markdown: markdown)
-        }
-        completeSuccessfulEditorCommit(exitAfter: exitAfter, rerender: true)
-    }
-
-    private func completeSuccessfulEditorCommit(exitAfter: Bool, rerender: Bool) {
-        if hasUnsavedEditorChanges {
-            if exitAfter {
-                pendingCommitShouldExit = true
-            }
-            finishEditorCommit(success: true)
-            return
-        }
-        guard exitAfter else {
-            finishEditorCommit(success: true)
-            return
-        }
-        exitEditMode(rerender: rerender) { [weak self] in
-            self?.finishEditorCommit(success: true)
-        }
-    }
-
-    private func finishEditorCommit(success: Bool) {
-        isEditorCommitInFlight = false
-        if pendingEditorCommitRequested {
-            pendingEditorCommitRequested = false
-            isPerformingAutomaticSave = false
-            performPendingEditorCommit()
-            return
-        }
-        if success,
-           pendingCommitShouldExit
-            || (hasUnsavedEditorChanges && !isPerformingAutomaticSave) {
-            startAutoSaveTimerIfNeeded()
-            performPendingEditorCommit()
-            return
-        }
-
-        if hasUnsavedEditorChanges {
-            startAutoSaveTimerIfNeeded()
-        } else {
-            stopAutoSaveTimer()
-        }
-
-        pendingEditorCommitRequested = false
-        pendingCommitShouldExit = false
-        let completions = pendingCommitCompletions
-        pendingCommitCompletions.removeAll()
-        for completion in completions {
-            completion(success)
-        }
-    }
-
-    private func exitEditMode(rerender: Bool,
-                              preserveUnsavedChanges: Bool = false,
-                              hidesAccessoryAfterFade: Bool = false,
-                              completion: @escaping () -> Void) {
-        guard let split = mainSplit else {
-            completion()
-            return
-        }
-        split.editorViewController?.contentDidChange = nil
-        split.editorViewController?.cancelRequested = nil
-        documentWindow.makeFirstResponder(nil)
-        let overlayHidden: (() -> Void)? = hidesAccessoryAfterFade
-            ? { [weak self] in self?.dismissEditChrome() }
-            : nil
-        split.exitEditMode(waitForPreviewRender: rerender,
-                           overlayHidden: overlayHidden) { [weak self] in
-            guard let self else {
-                completion()
-                return
-            }
-            if !preserveUnsavedChanges {
-                self.hasUnsavedEditorChanges = false
-            }
-            if self.editAccessory == nil {
-                self.updateEditToolbarItem()
-            }
-            if rerender, let url = self.currentFileURL, let markdown = self.currentMarkdown {
-                self.renderCurrentDocument(text: markdown, fileURL: url)
-            }
-            completion()
-        }
-    }
-
-    private func diskFileState(for url: URL?, expectedMarkdown: String?) -> DiskFileState {
-        guard let url, let expectedMarkdown else { return .unreadable }
-        do {
-            let diskMarkdown = try String(contentsOf: url, encoding: .utf8)
-            return diskMarkdown == expectedMarkdown ? .unchanged : .modified(diskMarkdown)
-        } catch {
-            return FileManager.default.fileExists(atPath: url.path) ? .unreadable : .missing
-        }
-    }
-
-    private func saveEditedMarkdown(_ text: String,
-                                    diskState: DiskFileState,
-                                    completion: @escaping (EditedMarkdownSaveResult) -> Void) {
-        guard let url = currentFileURL else {
-            completion(.cancelled)
-            return
-        }
-        if isPerformingAutomaticSave {
-            guard case .unchanged = diskState else {
-                completion(.cancelled)
-                return
-            }
-        }
-        switch diskState {
-        case .unchanged:
-            persistEditedMarkdown(text, to: url, completion: completion)
-        case let .modified(externalMarkdown):
-            presentExternalEditConflict(localMarkdown: text,
-                                        externalMarkdown: externalMarkdown,
-                                        fileURL: url,
-                                        completion: completion)
-        case .missing:
-            presentUnavailableFileConflict(localMarkdown: text,
-                                           fileURL: url,
-                                           reason: NSLocalizedString(
-                                               "The file was removed while it was open.",
-                                               comment: "Unavailable file conflict reason"
-                                           ),
-                                           overwriteTitle: NSLocalizedString(
-                                               "Recreate File",
-                                               comment: "Unavailable file conflict button"
-                                           ),
-                                           completion: completion)
-        case .unreadable:
-            presentUnavailableFileConflict(localMarkdown: text,
-                                           fileURL: url,
-                                           reason: NSLocalizedString(
-                                               "The file could not be read to verify that it is unchanged.",
-                                               comment: "Unavailable file conflict reason"
-                                           ),
-                                           overwriteTitle: NSLocalizedString(
-                                               "Save Anyway",
-                                               comment: "Unavailable file conflict button"
-                                           ),
-                                           completion: completion)
-        }
-    }
-
-    private func toggleTaskCheckbox(onLine sourceLine: Int, checked: Bool) {
-        guard !isEditing,
-              let baseline = currentMarkdown,
-              let updated = TaskCheckboxSource.settingChecked(
-                checked, onLine: sourceLine, in: baseline
-              ),
-              updated != baseline else {
-            rerenderCurrentPreview()
-            return
-        }
-
-        let diskState = diskFileState(for: currentFileURL, expectedMarkdown: baseline)
-        saveEditedMarkdown(updated, diskState: diskState) { [weak self] result in
-            guard let self else { return }
-            switch result {
-            case .saved:
-                self.currentMarkdown = updated
-                if let url = self.currentFileURL {
-                    self.markdownDocument?.replaceContents(markdown: updated, fileURL: url)
-                    self.renderCurrentDocument(text: updated, fileURL: url)
-                }
-            case let .reloaded(externalMarkdown):
-                self.currentMarkdown = externalMarkdown
-                if let url = self.currentFileURL {
-                    self.markdownDocument?.replaceContents(markdown: externalMarkdown, fileURL: url)
-                    self.renderCurrentDocument(text: externalMarkdown, fileURL: url)
-                }
-            case .cancelled:
-                self.rerenderCurrentPreview()
-            }
-        }
-    }
-
-    private func applyTableEdit(_ request: MarkdownTableEditRequest) {
-        guard !isEditing,
-              let baseline = currentMarkdown else {
-            rerenderCurrentPreview()
-            return
-        }
-        let updated = request.edits.reduce(Optional(baseline)) { markdown, edit in
-            markdown.flatMap {
-                MarkdownTableSource.applying(
-                    edit,
-                    fromLine: request.startLine,
-                    throughLine: request.endLine,
-                    in: $0
-                )
-            }
-        }
-        guard let updated,
-              updated != baseline else {
-            rerenderCurrentPreview()
-            return
-        }
-
-        let diskState = diskFileState(for: currentFileURL, expectedMarkdown: baseline)
-        let actionName = tableUndoActionName(for: request.edits)
-        saveEditedMarkdown(updated, diskState: diskState) { [weak self] result in
-            guard let self else { return }
-            switch result {
-            case .saved:
-                self.currentMarkdown = updated
-                if let url = self.currentFileURL {
-                    self.markdownDocument?.replaceContents(markdown: updated, fileURL: url)
-                    self.renderCurrentDocument(text: updated, fileURL: url)
-                    self.registerTableUndo(
-                        restoring: baseline,
-                        expectedCurrentMarkdown: updated,
-                        fileURL: url,
-                        actionName: actionName
-                    )
-                }
-            case let .reloaded(externalMarkdown):
-                self.tableUndoManager.removeAllActions()
-                self.currentMarkdown = externalMarkdown
-                if let url = self.currentFileURL {
-                    self.markdownDocument?.replaceContents(markdown: externalMarkdown, fileURL: url)
-                    self.renderCurrentDocument(text: externalMarkdown, fileURL: url)
-                }
-            case .cancelled:
-                self.rerenderCurrentPreview()
-            }
-        }
-    }
-
-    private func tableUndoActionName(for edits: [MarkdownTableEdit]) -> String {
-        for edit in edits.reversed() {
-            switch edit {
-            case .setCell:
-                return NSLocalizedString("Edit Table Cell", comment: "Table edit undo action")
-            case .insertRowBefore, .insertRowAfter:
-                return NSLocalizedString("Add Row", comment: "Table edit undo action")
-            case .deleteRow:
-                return NSLocalizedString("Delete Row", comment: "Table edit undo action")
-            case .insertColumnBefore, .insertColumnAfter:
-                return NSLocalizedString("Add Column", comment: "Table edit undo action")
-            case .deleteColumn:
-                return NSLocalizedString("Delete Column", comment: "Table edit undo action")
-            }
-        }
-        return NSLocalizedString("Edit Table", comment: "Table edit undo action")
-    }
-
-    private func registerTableUndo(restoring markdown: String,
-                                   expectedCurrentMarkdown: String,
-                                   fileURL: URL,
-                                   actionName: String) {
-        tableUndoManager.registerUndo(withTarget: self) { target in
-            target.restoreTableMarkdown(
-                markdown,
-                expectedCurrentMarkdown: expectedCurrentMarkdown,
-                fileURL: fileURL,
-                actionName: actionName
-            )
-        }
-        tableUndoManager.setActionName(actionName)
-    }
-
-    private func restoreTableMarkdown(_ markdown: String,
-                                      expectedCurrentMarkdown: String,
-                                      fileURL: URL,
-                                      actionName: String) {
-        guard !isEditing,
-              !isTableUndoSaveInFlight,
-              currentFileURL?.standardizedFileURL == fileURL.standardizedFileURL,
-              currentMarkdown == expectedCurrentMarkdown else {
-            tableUndoManager.removeAllActions()
-            return
-        }
-
-        // Register while UndoManager is actively undoing or redoing so AppKit
-        // puts this inverse operation on the opposite stack immediately. The
-        // file write can become asynchronous if sandbox permission is needed.
-        registerTableUndo(
-            restoring: expectedCurrentMarkdown,
-            expectedCurrentMarkdown: markdown,
-            fileURL: fileURL,
-            actionName: actionName
-        )
-        isTableUndoSaveInFlight = true
-        let diskState = diskFileState(for: currentFileURL,
-                                      expectedMarkdown: expectedCurrentMarkdown)
-        saveEditedMarkdown(markdown, diskState: diskState) { [weak self] result in
-            guard let self else { return }
-            self.isTableUndoSaveInFlight = false
-            switch result {
-            case .saved:
-                self.currentMarkdown = markdown
-                self.markdownDocument?.replaceContents(markdown: markdown, fileURL: fileURL)
-                self.renderCurrentDocument(text: markdown, fileURL: fileURL)
-            case let .reloaded(externalMarkdown):
-                self.tableUndoManager.removeAllActions()
-                self.currentMarkdown = externalMarkdown
-                self.markdownDocument?.replaceContents(
-                    markdown: externalMarkdown,
-                    fileURL: fileURL
-                )
-                self.renderCurrentDocument(text: externalMarkdown, fileURL: fileURL)
-            case .cancelled:
-                self.tableUndoManager.removeAllActions()
-                self.rerenderCurrentPreview()
-            }
-        }
-    }
-
-    private func rerenderCurrentPreview() {
-        guard let url = currentFileURL, let markdown = currentMarkdown else { return }
-        renderCurrentDocument(text: markdown, fileURL: url)
-    }
-
-    private func presentExternalEditConflict(
-        localMarkdown: String,
-        externalMarkdown: String,
-        fileURL: URL,
-        completion: @escaping (EditedMarkdownSaveResult) -> Void
-    ) {
-        let alert = NSAlert()
-        alert.alertStyle = .warning
-        alert.messageText = NSLocalizedString(
-            "This document changed on disk",
-            comment: "External edit conflict alert title"
-        )
-        alert.informativeText = String(
-            format: NSLocalizedString(
-                "Another app changed %@. Cancel keeps your changes unsaved. Choose which version to keep.",
-                comment: "External edit conflict alert message"
-            ),
-            fileURL.lastPathComponent
-        )
-        alert.addButton(withTitle: NSLocalizedString("Keep My Changes", comment: "Conflict alert button"))
-        alert.addButton(withTitle: NSLocalizedString("Reload from Disk", comment: "Conflict alert button"))
-        alert.addButton(withTitle: NSLocalizedString("Cancel", comment: "Alert button"))
-        alert.beginSheetModal(for: documentWindow) { [weak self] response in
-            guard let self else {
-                completion(.cancelled)
-                return
-            }
-            switch response {
-            case .alertFirstButtonReturn:
-                self.persistEditedMarkdown(localMarkdown, to: fileURL, completion: completion)
-            case .alertSecondButtonReturn:
-                // The sheet may remain open while another editor writes
-                // again. Reload the latest bytes instead of the snapshot
-                // captured when the conflict was first detected.
-                if let latestMarkdown = try? String(contentsOf: fileURL, encoding: .utf8) {
-                    completion(.reloaded(latestMarkdown))
-                } else {
-                    completion(.reloaded(externalMarkdown))
-                }
-            default:
-                completion(.cancelled)
-            }
-        }
-    }
-
-    private func presentUnavailableFileConflict(
-        localMarkdown: String,
-        fileURL: URL,
-        reason: String,
-        overwriteTitle: String,
-        completion: @escaping (EditedMarkdownSaveResult) -> Void
-    ) {
-        let alert = NSAlert()
-        alert.alertStyle = .warning
-        alert.messageText = NSLocalizedString(
-            "Unable to verify the document on disk",
-            comment: "Unavailable file conflict alert title"
-        )
-        alert.informativeText = String(
-            format: NSLocalizedString(
-                "%@ Cancel keeps your changes unsaved.",
-                comment: "Unavailable file conflict alert message"
-            ),
-            reason
-        )
-        alert.addButton(withTitle: overwriteTitle)
-        alert.addButton(withTitle: NSLocalizedString("Cancel", comment: "Alert button"))
-        alert.beginSheetModal(for: documentWindow) { [weak self] response in
-            guard let self, response == .alertFirstButtonReturn else {
-                completion(.cancelled)
-                return
-            }
-            self.persistEditedMarkdown(localMarkdown, to: fileURL, completion: completion)
-        }
-    }
-
-    private func persistEditedMarkdown(
-        _ text: String,
-        to url: URL,
-        completion: @escaping (EditedMarkdownSaveResult) -> Void
-    ) {
-        if write(text, to: url) {
-            completion(.saved)
-            return
-        }
-        guard !isPerformingAutomaticSave else {
-            completion(.cancelled)
-            return
-        }
-        // Sandbox denied the write — the file came in through the read-only
-        // filesystem exception (folder navigator) rather than a user-selected
-        // grant. A save panel pointed at the same file converts the user's
-        // confirmation into a read-write grant.
-        let panel = NSSavePanel()
-        panel.directoryURL = url.deletingLastPathComponent()
-        panel.nameFieldStringValue = url.lastPathComponent
-        panel.message = NSLocalizedString(
-            "Markdown Preview needs your permission to save this file.",
-            comment: "Save panel permission message"
-        )
-        panel.beginSheetModal(for: documentWindow) { [weak self] response in
-            guard let self, response == .OK, let chosen = panel.url else {
-                completion(.cancelled)
-                return
-            }
-            let ok = self.write(text, to: chosen)
-            if ok, chosen.standardizedFileURL != url.standardizedFileURL {
-                // Saved under a different name — follow the new file.
-                self.handleRename(to: chosen)
-            }
-            completion(ok ? .saved : .cancelled)
-        }
-    }
-
-    private func write(_ text: String, to url: URL) -> Bool {
-        // Atomic first (safe against partial writes); a file-scoped sandbox
-        // grant can deny the temp-file rename, so fall back to in-place.
-        do {
-            try text.write(to: url, atomically: true, encoding: .utf8)
-            return true
-        } catch {
-            do {
-                try text.write(to: url, atomically: false, encoding: .utf8)
-                return true
-            } catch {
-                return false
-            }
-        }
-    }
-
-    // MARK: - Open
-
-    /// Re-resolves the default hand-off app after Settings changes it.
-    func refreshOpenTargets() {
-        refreshOpenWithItem()
-        refreshOpenInLLMItem()
-        refreshOpenActionsItem()
-    }
-
-    private func makeOpenActionsItem() -> NSToolbarItem {
-        let item = NSMenuToolbarItem(itemIdentifier: .openActions)
-        let open = NSLocalizedString("Open", comment: "Open toolbar item label")
-        item.label = open
-        item.paletteLabel = open
-        item.toolTip = NSLocalizedString("Open document in another app", comment: "Open toolbar item tooltip")
-        item.target = self
-        item.action = #selector(openActionsPrimaryAction(_:))
-        item.showsIndicator = true
-        openActionsItem = item
-        refreshOpenActionsItem()
-        return item
-    }
-
-    private func refreshOpenActionsItem() {
-        let editors = currentFileURL.map { editorCandidates(for: $0) } ?? []
-        let defaultEditor = resolveDefaultEditor(among: editors)
-        let llmApps = llmCandidates()
-        let defaultLLM = resolveDefaultLLM(among: llmApps)
-        let defaultAction = resolveDefaultOpenAction(editors: editors,
-                                                     defaultEditor: defaultEditor,
-                                                     llmApps: llmApps,
-                                                     defaultLLM: defaultLLM)
-
-        let primaryTitle = openActionsTitle(for: defaultAction)
-
-        openActionsItem?.label = NSLocalizedString("Open", comment: "Open toolbar item label")
-        openActionsItem?.image = openActionsImage(for: defaultAction)
-        openActionsItem?.toolTip = primaryTitle ?? NSLocalizedString(
-            "Open document in another app",
-            comment: "Open toolbar item tooltip"
-        )
-        openActionsItem?.menu = buildOpenActionsMenu(editorCandidates: editors,
-                                                     llmCandidates: llmApps,
-                                                     defaultAction: defaultAction)
-    }
-
-    private func openActionsTitle(for selection: OpenActionSelection?) -> String? {
-        switch selection {
-        case .editor(let editor):
-            return localizedOpenIn(displayName(for: editor.url))
-        case .llm(let candidate):
-            return localizedOpenIn(candidate.target.title)
-        case nil:
-            return nil
-        }
-    }
-
-    private func localizedOpenIn(_ applicationName: String) -> String {
-        String(
-            format: NSLocalizedString("Open in %@", comment: "Open toolbar item for a named application"),
-            applicationName
-        )
-    }
-
-    private func openActionsImage(for selection: OpenActionSelection?) -> NSImage {
-        switch selection {
-        case .editor(let editor):
-            let editorURL = editor.url
-            return openWithImage(for: editorURL)
-        case .llm(let candidate):
-            return openInLLMImage(for: candidate)
-        case nil:
-            return openWithImage(for: nil)
-        }
-    }
-
-    private func resolveDefaultOpenAction(editors: [EditorCandidate],
-                                          defaultEditor: EditorCandidate?,
-                                          llmApps: [LLMCandidate],
-                                          defaultLLM: LLMCandidate?) -> OpenActionSelection? {
-        OpenTargetCatalog.resolveDefaultOpenAction(editors: editors,
-                                                   defaultEditor: defaultEditor,
-                                                   llmApps: llmApps,
-                                                   defaultLLM: defaultLLM)
-    }
-
-    private func buildOpenActionsMenu(editorCandidates: [EditorCandidate],
-                                      llmCandidates: [LLMCandidate],
-                                      defaultAction: OpenActionSelection?) -> NSMenu {
-        let menu = NSMenu()
-
-        guard currentFileURL != nil else {
-            menu.addItem(disabledItem(NSLocalizedString("No document open", comment: "Open menu empty state")))
-            return menu
-        }
-
-        if editorCandidates.isEmpty && llmCandidates.isEmpty {
-            menu.addItem(disabledItem(NSLocalizedString("No apps available", comment: "Open menu empty state")))
-            return menu
-        }
-
-        if !llmCandidates.isEmpty {
-            let header = NSMenuItem()
-            header.title = NSLocalizedString("AI Apps", comment: "Open menu section header")
-            header.isEnabled = false
-            menu.addItem(header)
-
-            for candidate in llmCandidates {
-                let item = NSMenuItem(
-                    title: candidate.target.title,
-                    action: #selector(pickLLMTarget(_:)),
-                    keyEquivalent: ""
-                )
-                item.target = self
-                item.representedObject = candidate.target.id
-                let icon = NSWorkspace.shared.icon(forFile: candidate.appURL.path)
-                icon.size = NSSize(width: 16, height: 16)
-                item.image = icon
-                if case .llm(let selectedLLM) = defaultAction,
-                   candidate.target.id == selectedLLM.target.id {
-                    item.state = .on
-                }
-                menu.addItem(item)
-            }
-        }
-
-        if !editorCandidates.isEmpty {
-            if !llmCandidates.isEmpty {
-                menu.addItem(.separator())
-            }
-
-            let header = NSMenuItem()
-            header.title = NSLocalizedString("Editors", comment: "Open menu section header")
-            header.isEnabled = false
-            menu.addItem(header)
-
-            for candidate in editorCandidates {
-                let item = NSMenuItem(
-                    title: displayName(for: candidate.url),
-                    action: #selector(pickEditor(_:)),
-                    keyEquivalent: ""
-                )
-                let icon = NSWorkspace.shared.icon(forFile: candidate.url.path)
-                icon.size = NSSize(width: 16, height: 16)
-                item.image = icon
-                item.target = self
-                item.representedObject = candidate
-                if case .editor(let selectedEditor) = defaultAction,
-                   sameEditor(candidate, selectedEditor) {
-                    item.state = .on
-                }
-                menu.addItem(item)
-            }
-        }
-
-        return menu
-    }
-
-    @objc private func openActionsPrimaryAction(_ sender: Any?) {
-        guard let fileURL = currentFileURL else { return }
-        let editors = editorCandidates(for: fileURL)
-        let defaultEditor = resolveDefaultEditor(among: editors)
-        let llmApps = llmCandidates()
-        let defaultLLM = resolveDefaultLLM(among: llmApps)
-        guard let defaultAction = resolveDefaultOpenAction(editors: editors,
-                                                           defaultEditor: defaultEditor,
-                                                           llmApps: llmApps,
-                                                           defaultLLM: defaultLLM) else {
-            NSSound.beep()
-            return
-        }
-
-        switch defaultAction {
-        case .editor(let editor):
-            launch(fileURL, with: editor.url)
-        case .llm(let target):
-            openInLLM(target, fileURL: fileURL)
-        }
-    }
-
-    // MARK: - Open in LLM
-
-    private static let llmDeepLinkCharacterLimit = 12_000
-    private static let claudeColdLaunchDeepLinkDelay: TimeInterval = 1.25
-    private static let chatGPTColdLaunchFileOpenDelay: TimeInterval = 1.25
-
-    private var hasLLMTargetsAvailable: Bool {
-        !llmCandidates().isEmpty
-    }
-
-    private func makeOpenInLLMItem() -> NSToolbarItem {
-        let item = NSMenuToolbarItem(itemIdentifier: .openInLLM)
-        let openInLLM = NSLocalizedString("Open in LLM", comment: "Open in LLM toolbar item label")
-        item.label = openInLLM
-        item.paletteLabel = openInLLM
-        item.toolTip = NSLocalizedString("Open document in an LLM app", comment: "Open in LLM toolbar item tooltip")
-        item.target = self
-        item.action = #selector(openInLLMPrimaryAction(_:))
-        item.showsIndicator = true
-        openInLLMItem = item
-        refreshOpenInLLMItem()
-        return item
-    }
-
-    private func refreshOpenInLLMItem() {
-        let candidates = llmCandidates()
-        guard !candidates.isEmpty else {
-            removeOpenInLLMToolbarItem()
-            return
-        }
-        let resolvedDefault = resolveDefaultLLM(among: candidates)
-        let openInTitle = resolvedDefault.map { localizedOpenIn($0.target.title) }
-        openInLLMItem?.label = openInTitle ?? NSLocalizedString("Open in LLM", comment: "Open in LLM toolbar item label")
-        openInLLMItem?.image = openInLLMImage(for: resolvedDefault)
-        openInLLMItem?.toolTip = openInTitle ?? NSLocalizedString(
-            "Open document in an LLM app",
-            comment: "Open in LLM toolbar item tooltip"
-        )
-        openInLLMItem?.menu = buildOpenInLLMMenu(candidates: candidates,
-                                                 defaultTarget: resolvedDefault)
-    }
-
-    private func removeOpenInLLMToolbarItem() {
-        guard let toolbar = documentWindow.toolbar,
-              let index = toolbar.items.firstIndex(where: { $0.itemIdentifier == .openInLLM }) else {
-            return
-        }
-        toolbar.removeItem(at: index)
-    }
-
-    private func openInLLMImage(for candidate: LLMCandidate?) -> NSImage {
-        if let appURL = candidate?.appURL {
-            let icon = NSWorkspace.shared.icon(forFile: appURL.path)
-            icon.size = NSSize(width: 20, height: 20)
-            return icon
-        }
-        return NSImage(systemSymbolName: "sparkles",
-                       accessibilityDescription: NSLocalizedString("Open in LLM", comment: "Open in LLM toolbar image")) ?? NSImage()
-    }
-
-    private func llmCandidates() -> [LLMCandidate] {
-        OpenTargetCatalog.llmCandidates()
-    }
-
-    private func resolveDefaultLLM(among candidates: [LLMCandidate]) -> LLMCandidate? {
-        OpenTargetCatalog.resolveDefaultLLM(among: candidates)
-    }
-
-    private func buildOpenInLLMMenu(candidates: [LLMCandidate],
-                                    defaultTarget: LLMCandidate?) -> NSMenu {
-        let menu = NSMenu()
-
-        guard currentFileURL != nil else {
-            menu.addItem(disabledItem(NSLocalizedString("No document open", comment: "Open menu empty state")))
-            return menu
-        }
-        guard !candidates.isEmpty else {
-            menu.addItem(disabledItem(NSLocalizedString("No LLM apps available", comment: "Open menu empty state")))
-            return menu
-        }
-
-        let header = NSMenuItem()
-        header.title = NSLocalizedString("Open in LLM…", comment: "Open in LLM menu header")
-        header.isEnabled = false
-        menu.addItem(header)
-
-        for candidate in candidates {
-            let item = NSMenuItem(
-                title: candidate.target.title,
-                action: #selector(pickLLMTarget(_:)),
-                keyEquivalent: ""
-            )
-            item.target = self
-            item.representedObject = candidate.target.id
-            let icon = NSWorkspace.shared.icon(forFile: candidate.appURL.path)
-            icon.size = NSSize(width: 16, height: 16)
-            item.image = icon
-            if let defaultTarget, candidate.target.id == defaultTarget.target.id {
-                item.state = .on
-            }
-            menu.addItem(item)
-        }
-        return menu
-    }
-
-    @objc private func openInLLMPrimaryAction(_ sender: Any?) {
-        guard let fileURL = currentFileURL else { return }
-        let candidates = llmCandidates()
-        guard let target = resolveDefaultLLM(among: candidates) else {
-            NSSound.beep()
-            return
-        }
-        openInLLM(target, fileURL: fileURL)
-    }
-
-    @objc private func pickLLMTarget(_ sender: NSMenuItem) {
-        guard let targetID = sender.representedObject as? String,
-              let candidate = llmCandidates().first(where: { $0.target.id == targetID }),
-              let fileURL = currentFileURL else { return }
-        OpenTargetCatalog.setDefaultLLM(candidate)
-        refreshOpenInLLMItem()
-        refreshOpenActionsItem()
-        openInLLM(candidate, fileURL: fileURL)
-    }
-
-    private func openInLLM(_ candidate: LLMCandidate, fileURL: URL) {
-        let folderURL = fileURL.deletingLastPathComponent()
-
-        switch candidate.target.handoff {
-        case .codexDesktop:
-            let prompt = llmPathPrompt(for: fileURL)
-            if let url = codexDeepLink(prompt: prompt, folderURL: folderURL) {
-                NSWorkspace.shared.open(url)
-            } else {
-                copyPromptAndOpen(candidate: candidate, prompt: prompt)
-            }
-        case .claudeCodeDesktop:
-            let prompt = llmEmbeddedMarkdownPrompt(for: fileURL)
-            if prompt.count <= Self.llmDeepLinkCharacterLimit,
-               let url = claudeCodeDeepLink(prompt: prompt, folderURL: folderURL) {
-                openDeepLink(url, afterLaunchingIfNeeded: candidate, delay: Self.claudeColdLaunchDeepLinkDelay)
-            } else {
-                copyPromptAndOpen(candidate: candidate, prompt: prompt)
-            }
-        case .chatGPTDocumentOpen:
-            openDocumentInChatGPT(fileURL, candidate: candidate)
-        default:
-            let prompt = llmPathPrompt(for: fileURL)
-            copyPromptAndOpen(candidate: candidate, prompt: prompt)
-        }
-    }
-
-    private func openDeepLink(_ url: URL, afterLaunchingIfNeeded candidate: LLMCandidate, delay: TimeInterval) {
-        guard !isRunning(candidate) else {
-            NSWorkspace.shared.open(url)
-            return
-        }
-
-        let configuration = NSWorkspace.OpenConfiguration()
-        configuration.activates = true
-        NSWorkspace.shared.openApplication(
-            at: candidate.appURL,
-            configuration: configuration
-        ) { _, error in
-            if error != nil {
-                NSSound.beep()
-                return
-            }
-            DispatchQueue.main.asyncAfter(deadline: .now() + delay) {
-                NSWorkspace.shared.open(url)
-            }
-        }
-    }
-
-    private func openDocumentInChatGPT(_ fileURL: URL, candidate: LLMCandidate) {
-        if isRunning(candidate) {
-            sendDocumentOpenEventToChatGPT(fileURL, candidate: candidate)
-            return
-        }
-
-        let delay = Self.chatGPTColdLaunchFileOpenDelay
-        let configuration = NSWorkspace.OpenConfiguration()
-        configuration.activates = true
-        NSWorkspace.shared.openApplication(
-            at: candidate.appURL,
-            configuration: configuration
-        ) { [weak self] _, error in
-            if error != nil {
-                NSSound.beep()
-                return
-            }
-            DispatchQueue.main.asyncAfter(deadline: .now() + delay) {
-                self?.sendDocumentOpenEventToChatGPT(fileURL, candidate: candidate)
-            }
-        }
-    }
-
-    private func isRunning(_ candidate: LLMCandidate) -> Bool {
-        candidate.target.bundleIDs.contains { bundleID in
-            !NSRunningApplication.runningApplications(withBundleIdentifier: bundleID).isEmpty
-        }
-    }
-
-    private func sendDocumentOpenEventToChatGPT(_ fileURL: URL, candidate: LLMCandidate) {
-        let configuration = NSWorkspace.OpenConfiguration()
-        configuration.activates = true
-        NSWorkspace.shared.open(
-            [fileURL],
-            withApplicationAt: candidate.appURL,
-            configuration: configuration
-        ) { _, error in
-            if error != nil {
-                NSSound.beep()
-            }
-        }
-    }
-
-    private func copyPromptAndOpen(candidate: LLMCandidate, prompt: String) {
-        let pasteboard = NSPasteboard.general
-        pasteboard.clearContents()
-        pasteboard.setString(prompt, forType: .string)
-
-        let configuration = NSWorkspace.OpenConfiguration()
-        configuration.activates = true
-        NSWorkspace.shared.openApplication(
-            at: candidate.appURL,
-            configuration: configuration
-        ) { _, _ in }
-    }
-
-    private func codexDeepLink(prompt: String, folderURL: URL) -> URL? {
-        var components = URLComponents()
-        components.scheme = "codex"
-        components.host = "new"
-        components.queryItems = [
-            URLQueryItem(name: "prompt", value: prompt),
-            URLQueryItem(name: "path", value: folderURL.path)
-        ]
-        return components.url
-    }
-
-    private func claudeCodeDeepLink(prompt: String, folderURL: URL) -> URL? {
-        var components = URLComponents()
-        components.scheme = "claude"
-        components.host = "code"
-        components.path = "/new"
-        components.queryItems = [
-            URLQueryItem(name: "q", value: prompt),
-            URLQueryItem(name: "folder", value: folderURL.path)
-        ]
-        return components.url
-    }
-
-    private func llmPathPrompt(for fileURL: URL) -> String {
-        """
-        Open this Markdown file and use it as the working context:
-        \(fileURL.path)
-        """
-    }
-
-    private func llmEmbeddedMarkdownPrompt(for fileURL: URL) -> String {
-        guard let markdown = currentMarkdown
-                ?? (try? String(contentsOf: fileURL, encoding: .utf8)),
-              !markdown.isEmpty else {
-            return llmPathPrompt(for: fileURL)
-        }
-
-        return """
-        Use this Markdown document as the working context.
-
-        The local path is included only as a reference. Do not rely on opening it to read the document contents.
-
-        File name: \(fileURL.lastPathComponent)
-        Local path: \(fileURL.path)
-
-        Markdown content:
-        ````markdown
-        \(markdown)
-        ````
-        """
-    }
-
-    private func makeZoomItem() -> NSToolbarItemGroup {
-        let zoomOut = NSLocalizedString("Zoom Out", comment: "Zoom toolbar button")
-        let zoomIn = NSLocalizedString("Zoom In", comment: "Zoom toolbar button")
-        let zoom = NSLocalizedString("Zoom", comment: "Zoom toolbar item label")
-        let smaller = NSImage(systemSymbolName: "textformat.size.smaller",
-                              accessibilityDescription: zoomOut) ?? NSImage()
-        let larger = NSImage(systemSymbolName: "textformat.size.larger",
-                             accessibilityDescription: zoomIn) ?? NSImage()
-        let group = NSToolbarItemGroup(
-            itemIdentifier: .zoom,
-            images: [smaller, larger],
-            selectionMode: .momentary,
-            labels: [zoomOut, zoomIn],
-            target: self,
-            action: #selector(zoomSegmentAction(_:))
-        )
-        group.label = zoom
-        group.paletteLabel = zoom
-        group.toolTip = zoom
-        for (subitem, tooltip) in zip(group.subitems, [zoomOut, zoomIn]) {
-            subitem.toolTip = tooltip
-        }
-        // .expanded keeps the two-segment "A A" pair visible like Books / Reader,
-        // instead of collapsing into a single button + menu when space is tight.
-        group.controlRepresentation = .expanded
-        if let segmented = group.view as? NSSegmentedControl {
-            segmented.setToolTip(zoomOut, forSegment: 0)
-            segmented.setToolTip(zoomIn, forSegment: 1)
-        }
-        return group
-    }
-
-    @objc private func zoomSegmentAction(_ sender: NSToolbarItemGroup) {
-        guard let split = documentWindow.contentViewController as? MainSplitViewController else { return }
-        switch sender.selectedIndex {
-        case 0: split.zoomOutDocument(sender)
-        case 1: split.zoomInDocument(sender)
-        default: break
-        }
-    }
-
-    private func inspectorImage() -> NSImage {
-        let image = NSImage(systemSymbolName: "info.circle",
-                            accessibilityDescription: NSLocalizedString("Inspector", comment: "Inspector toolbar image")) ?? NSImage()
-        image.isTemplate = true
-        return image
-    }
-
-    @objc private func toggleInspectorAction(_ sender: Any) {
-        let isVisible = (documentWindow.contentViewController as? MainSplitViewController)?
-            .toggleInspector() ?? false
-        setInspectorToggleSelected(isVisible)
-    }
-
-    private func refreshInspectorToggleItem() {
-        let isVisible = (documentWindow.contentViewController as? MainSplitViewController)?
-            .isInspectorVisible ?? false
-        setInspectorToggleSelected(isVisible)
-    }
-
-    private func setInspectorToggleSelected(_ isSelected: Bool) {
-        isInspectorToggleSelected = isSelected
-        inspectorButton?.state = isSelected ? .on : .off
-    }
-
-    func items(for pickerToolbarItem: NSSharingServicePickerToolbarItem) -> [Any] {
-        guard let currentMarkdown else { return [] }
-        return [currentMarkdown]
-    }
-
-    private func makeSearchItem() -> NSToolbarItem {
-        let item = NSSearchToolbarItem(itemIdentifier: .search)
-        let searchInDocument = NSLocalizedString(
-            "Search in Document",
-            comment: "Toolbar search field tooltip and placeholder"
-        )
-        item.label = NSLocalizedString("Search", comment: "Toolbar search item label")
-        item.toolTip = searchInDocument
-        item.preferredWidthForSearchField = 320
-        item.searchField.placeholderString = searchInDocument
-        item.searchField.sendsSearchStringImmediately = true
-        item.searchField.target = self
-        item.searchField.action = #selector(searchFieldDidChange(_:))
-        item.searchField.delegate = self
-        searchField = item.searchField
-        return item
-    }
-
-    @objc private func searchFieldDidChange(_ sender: NSSearchField) {
-        // Coalesce per-keystroke finds — running the full DOM rewrite + JS
-        // round-trip on every char is the dominant stall source on big docs.
-        // Empty queries (e.g. user cleared the field) bypass the debounce so
-        // the highlight teardown happens immediately.
-        let query = sender.stringValue
-        pendingFindWork?.cancel()
-        if query.isEmpty {
-            pendingFindWork = nil
-            runFind(query: query)
-            return
-        }
-        let work = DispatchWorkItem { [weak self] in
-            self?.runFind(query: query)
-        }
-        pendingFindWork = work
-        DispatchQueue.main.asyncAfter(
-            deadline: .now() + Self.findDebounceDelay, execute: work
-        )
-    }
-
-    private func runFind(query: String, backwards: Bool = false) {
-        // Explicit nav (Enter / prev / next / mode change) flushes any pending
-        // debounce so the user navigates the freshest results.
-        pendingFindWork?.cancel()
-        pendingFindWork = nil
-        (documentWindow.contentViewController as? MainSplitViewController)?
-            .find(query, backwards: backwards, mode: searchMode) { [weak self] result in
-                self?.applyFindResult(result, query: query)
-            }
-    }
-
-    func control(_ control: NSControl,
-                 textView: NSTextView,
-                 doCommandBy commandSelector: Selector) -> Bool {
-        guard control === searchField,
-              commandSelector == #selector(NSResponder.insertNewline(_:)) else {
-            return false
-        }
-        let backwards = NSEvent.modifierFlags.contains(.shift)
-        findFromToolbar(backwards: backwards)
-        return true
-    }
-
-    private func applyFindResult(_ result: FindResult, query: String) {
-        if query.isEmpty {
-            setFindBarVisible(false)
-            return
-        }
-        findBar?.update(matchCount: result.total, currentIndex: result.index)
-        setFindBarVisible(true)
-    }
-
-    private func setFindBarVisible(_ visible: Bool) {
-        guard let accessory = findBarAccessory, accessory.isHidden == visible else { return }
-        accessory.isHidden = !visible
-        if #available(macOS 26.1, *) {
-            // Visible bar always gets the hard backdrop; hidden, the
-            // preference must not leak a backdrop over a themed titlebar.
-            accessory.preferredScrollEdgeEffectStyle =
-                visible || !activeSchemeThemed ? .hard : .automatic
-        }
-    }
-
-    private func installFindBar() {
-        let bar = FindBar(
-            frame: NSRect(x: 0, y: 0, width: 600, height: FindBar.preferredHeight)
-        )
-        bar.autoresizingMask = [.width]
-        bar.onPrevious = { [weak self] in self?.findFromToolbar(backwards: true) }
-        bar.onNext = { [weak self] in self?.findFromToolbar(backwards: false) }
-        bar.onDone = { [weak self] in self?.dismissFindBar() }
-        bar.onModeChanged = { [weak self] mode in self?.searchModeDidChange(mode) }
-        self.findBar = bar
-        // No .hard here: a hidden accessory's scroll-edge preference still
-        // applies, painting an opaque backdrop over a themed window background.
-        // setFindBarVisible flips it while the bar is actually shown.
-        self.findBarAccessory = addBottomTitlebarAccessory(bar)
-    }
-
-    private func dismissFindBar() {
-        searchField?.stringValue = ""
-        if let editor = searchField?.currentEditor(),
-           documentWindow.firstResponder === editor {
-            documentWindow.makeFirstResponder(nil)
-        }
-        runFind(query: "")
-    }
-
-    @IBAction func performFindPanelAction(_ sender: Any?) {
-        handleFindAction(sender)
-    }
-
-    @IBAction override func performTextFinderAction(_ sender: Any?) {
-        handleFindAction(sender)
-    }
-
-    func handleFindAction(_ sender: Any?) {
-        let tag = (sender as? NSValidatedUserInterfaceItem)?.tag ?? 1
-        switch tag {
-        case NSTextFinder.Action.nextMatch.rawValue:
-            findFromToolbar(backwards: false)
-        case NSTextFinder.Action.previousMatch.rawValue:
-            findFromToolbar(backwards: true)
-        default:
-            focusToolbarSearch()
-        }
-    }
-
-    private func findFromToolbar(backwards: Bool) {
-        let query = searchField?.stringValue
-            ?? NSPasteboard(name: .find).string(forType: .string)
-            ?? ""
-        guard !query.isEmpty else {
-            focusToolbarSearch()
-            return
-        }
-        runFind(query: query, backwards: backwards)
-    }
-
-    private func searchModeDidChange(_ mode: SearchMode) {
-        guard mode != searchMode else { return }
-        searchMode = mode
-        guard findBarAccessory?.isHidden == false,
-              let query = searchField?.stringValue, !query.isEmpty else { return }
-        runFind(query: query)
-    }
-
-    private func focusToolbarSearch() {
-        guard let searchField else { return }
-        documentWindow.makeFirstResponder(searchField)
-        searchField.selectText(nil)
-    }
-
-    // MARK: - Open With
-
-    private static let markdownFileExtensions = ["md", "markdown", "mdown", "txt"]
-
-    private func makeOpenWithItem() -> NSToolbarItem {
-        let item = NSMenuToolbarItem(itemIdentifier: .openWith)
-        let openWith = NSLocalizedString("Open With", comment: "Open With toolbar item label")
-        item.label = openWith
-        item.paletteLabel = openWith
-        item.toolTip = NSLocalizedString("Open in another editor", comment: "Open With toolbar item tooltip")
-        item.target = self
-        item.action = #selector(openWithPrimaryAction(_:))
-        item.showsIndicator = true
-        openWithItem = item
-        refreshOpenWithItem()
-        return item
-    }
-
-    private func refreshOpenWithItem() {
-        let candidates = currentFileURL.map { editorCandidates(for: $0) } ?? []
-        let resolvedDefault = resolveDefaultEditor(among: candidates)
-        let openInTitle = resolvedDefault.map { localizedOpenIn(displayName(for: $0.url)) }
-        openWithItem?.label = openInTitle ?? NSLocalizedString("Open With", comment: "Open With toolbar item label")
-        openWithItem?.image = openWithImage(for: resolvedDefault?.url)
-        openWithItem?.toolTip = openInTitle ?? NSLocalizedString(
-            "Open in another editor",
-            comment: "Open With toolbar item tooltip"
-        )
-        openWithItem?.menu = buildOpenWithMenu(candidates: candidates,
-                                               defaultEditor: resolvedDefault)
-    }
-
-    private func openWithImage(for url: URL?) -> NSImage {
-        if let url {
-            let icon = NSWorkspace.shared.icon(forFile: url.path)
-            icon.size = NSSize(width: 20, height: 20)
-            return icon
-        }
-        return NSImage(systemSymbolName: "highlighter",
-                       accessibilityDescription: NSLocalizedString("Open With", comment: "Open With toolbar image")) ?? NSImage()
-    }
-
-    @objc private func openWithPrimaryAction(_ sender: Any?) {
-        guard let fileURL = currentFileURL else { return }
-        let candidates = editorCandidates(for: fileURL)
-        if let editor = resolveDefaultEditor(among: candidates) {
-            launch(fileURL, with: editor.url)
-        }
-    }
-
-    private func editorCandidates(for fileURL: URL) -> [EditorCandidate] {
-        OpenTargetCatalog.editorCandidates(for: fileURL)
-    }
-
-    private func resolveDefaultEditor(among candidates: [EditorCandidate]) -> EditorCandidate? {
-        OpenTargetCatalog.resolveDefaultEditor(among: candidates)
-    }
-
-    private func buildOpenWithMenu(candidates: [EditorCandidate],
-                                   defaultEditor: EditorCandidate?) -> NSMenu {
-        let menu = NSMenu()
-
-        guard currentFileURL != nil else {
-            menu.addItem(disabledItem(NSLocalizedString("No document open", comment: "Open menu empty state")))
-            return menu
-        }
-        guard !candidates.isEmpty else {
-            menu.addItem(disabledItem(NSLocalizedString("No editors available", comment: "Open menu empty state")))
-            return menu
-        }
-
-        let header = NSMenuItem()
-        header.title = NSLocalizedString("Open with…", comment: "Open With menu header")
-        header.isEnabled = false
-        menu.addItem(header)
-
-        for candidate in candidates {
-            let item = NSMenuItem(
-                title: displayName(for: candidate.url),
-                action: #selector(pickEditor(_:)),
-                keyEquivalent: ""
-            )
-            let icon = NSWorkspace.shared.icon(forFile: candidate.url.path)
-            icon.size = NSSize(width: 16, height: 16)
-            item.image = icon
-            item.target = self
-            item.representedObject = candidate
-            if let defaultEditor, sameEditor(candidate, defaultEditor) {
-                item.state = .on
-            }
-            menu.addItem(item)
-        }
-        return menu
-    }
-
-    private func displayName(for appURL: URL) -> String {
-        OpenTargetCatalog.displayName(for: appURL)
-    }
-
-    private func sameEditor(_ lhs: EditorCandidate, _ rhs: EditorCandidate) -> Bool {
-        OpenTargetCatalog.sameEditor(lhs, rhs)
-    }
-
-    private func sameApplication(_ lhs: URL, _ rhs: URL) -> Bool {
-        OpenTargetCatalog.sameApplication(lhs, rhs)
-    }
-
-    private func canonicalAppURL(_ url: URL) -> URL {
-        OpenTargetCatalog.canonicalAppURL(url)
-    }
-
-    private func disabledItem(_ title: String) -> NSMenuItem {
-        let item = NSMenuItem(title: title, action: nil, keyEquivalent: "")
-        item.isEnabled = false
-        return item
-    }
-
-    @objc private func pickEditor(_ sender: NSMenuItem) {
-        guard let candidate = sender.representedObject as? EditorCandidate,
-              let fileURL = currentFileURL else { return }
-        OpenTargetCatalog.setDefaultEditor(candidate)
-        refreshOpenWithItem()
-        refreshOpenActionsItem()
-        launch(fileURL, with: candidate.url)
-    }
-
-    private func launch(_ fileURL: URL, with appURL: URL) {
-        NSWorkspace.shared.open(
-            [fileURL],
-            withApplicationAt: appURL,
-            configuration: NSWorkspace.OpenConfiguration()
-        ) { _, _ in }
-    }
-
-    private struct ContextOpenPayload {
-        let fileURL: URL
-        let appURL: URL
-    }
-
-    func openInNewTab(_ fileURL: URL) {
-        Self.markNextWindowAsTab()
-        openDocumentWindow(for: fileURL) {
-            // If the document was already open, no window was created and
-            // the override wasn't consumed — don't let it leak to the next one.
-            Self.nextWindowRequestsTab = false
-        }
-    }
-
-    func openInNewWindow(_ fileURL: URL) {
-        Self.markNextWindowAsSeparate()
-        openDocumentWindow(for: fileURL) {
-            // If the document was already open, no window was created and
-            // the override wasn't consumed — don't let it leak to the next one.
-            Self.nextWindowDeclinesTabbing = false
-        }
-    }
-
-    private func openDocumentWindow(for fileURL: URL, completion: (() -> Void)? = nil) {
-        NSDocumentController.shared.openDocument(withContentsOf: fileURL,
-                                                 display: true) { [weak self] _, _, error in
-            completion?()
-            guard let self, let error else { return }
-            NSAlert(error: error).beginSheetModal(for: self.documentWindow)
-        }
-    }
-
-    /// Backs File > New Tab. Deliberately NOT the NSResponder
-    /// `newWindowForTab(_:)` override: responding to that selector is what
-    /// makes AppKit show the "+" button in the tab bar, and the app hides
-    /// that button. There is no untitled-document concept here, so prompt
-    /// for a file and open it as a tab — an explicit tab request, unlike ⌘O.
-    @objc func newDocumentTab(_ sender: Any?) {
-        promptForDocument(openAsTab: true)
-    }
-
-    func openFolder(_ folderURL: URL) {
-        let folderURL = folderURL.standardizedFileURL
-        if currentFileURL == nil {
-            documentWindow.title = folderURL.lastPathComponent
-            updateWindowSubtitle()
-        }
-        (documentWindow.contentViewController as? MainSplitViewController)?
-            .openFolder(folderURL, selectedFileURL: currentFileURL)
-        documentWindow.makeKeyAndOrderFront(nil)
-        NSApp.activate()
-        syncSidebarMenuState()
-    }
-
-    func contextMenuEditorItems(for fileURL: URL) -> [NSMenuItem] {
-        let candidates = editorCandidates(for: fileURL)
-        let defaultEditor = resolveDefaultEditor(among: candidates)
-
-        var items: [NSMenuItem] = []
-
-        let externalItem = NSMenuItem(
-            title: NSLocalizedString("Open with External Editor", comment: "Context menu item"),
-            action: #selector(contextLaunchEditor(_:)),
-            keyEquivalent: ""
-        )
-        externalItem.image = NSImage(systemSymbolName: "arrow.up.right.square",
-                                     accessibilityDescription: nil)
-        if let defaultEditor {
-            externalItem.target = self
-            externalItem.representedObject = ContextOpenPayload(fileURL: fileURL, appURL: defaultEditor.url)
-            externalItem.toolTip = localizedOpenIn(displayName(for: defaultEditor.url))
-        } else {
-            externalItem.isEnabled = false
-        }
-        items.append(externalItem)
-
-        let openAs = NSMenuItem(
-            title: NSLocalizedString("Open As", comment: "Context menu submenu"),
-            action: nil,
-            keyEquivalent: ""
-        )
-        let submenu = NSMenu()
-        if candidates.isEmpty {
-            submenu.addItem(disabledItem(NSLocalizedString("No editors available", comment: "Open As empty state")))
-        } else {
-            for candidate in candidates {
-                let item = NSMenuItem(
-                    title: displayName(for: candidate.url),
-                    action: #selector(contextLaunchEditor(_:)),
-                    keyEquivalent: ""
-                )
-                item.target = self
-                item.representedObject = ContextOpenPayload(fileURL: fileURL, appURL: candidate.url)
-                let icon = NSWorkspace.shared.icon(forFile: candidate.url.path)
-                icon.size = NSSize(width: 16, height: 16)
-                item.image = icon
-                if let defaultEditor, sameEditor(candidate, defaultEditor) {
-                    item.state = .on
-                }
-                submenu.addItem(item)
-            }
-        }
-        openAs.submenu = submenu
-        items.append(openAs)
-
-        return items
-    }
-
-    @objc private func contextLaunchEditor(_ sender: NSMenuItem) {
-        guard let payload = sender.representedObject as? ContextOpenPayload else { return }
-        launch(payload.fileURL, with: payload.appURL)
-    }
-
-    @IBAction func openDocument(_ sender: Any?) {
-        promptForDocument(openAsTab: false)
-    }
-
-    private func promptForDocument(openAsTab: Bool) {
-        let panel = makeOpenPanel()
-        panel.beginSheetModal(for: documentWindow) { [weak self] response in
-            guard let self, response == .OK, let url = panel.url else { return }
-            if url.isExistingDirectory {
-                self.openFolder(url)
-                return
-            }
-            if openAsTab {
-                self.openInNewTab(url)
-            } else {
-                // Plain open: tab placement follows the system
-                // "Prefer tabs" setting via attachToExistingTabGroupIfNeeded.
-                self.openDocumentWindow(for: url)
-            }
-        }
-    }
-
-    private func makeOpenPanel() -> NSOpenPanel {
-        let panel = NSOpenPanel()
-        panel.allowsMultipleSelection = false
-        panel.canChooseDirectories = true
-        panel.canChooseFiles = true
-        panel.message = NSLocalizedString(
-            "Choose a Markdown file or folder",
-            comment: "Open panel prompt"
-        )
-        panel.allowedContentTypes = Self.markdownFileExtensions
-            .compactMap { UTType(filenameExtension: $0) }
-        return panel
-    }
-
-    private func loadFile(at url: URL, silentOnFailure: Bool = false) {
-        Task { @concurrent [weak self] in
-            do {
-                let text = try String(contentsOf: url, encoding: .utf8)
-                await self?.applyLoadedMarkdown(text, fileURL: url)
-            } catch {
-                // Wrap as NSError (Sendable) so the original presentation —
-                // localizedDescription + recovery suggestion — survives the
-                // hop back to MainActor.
-                let nsError = error as NSError
-                await self?.applyLoadFailure(error: nsError,
-                                             fileURL: url,
-                                             silentOnFailure: silentOnFailure)
-            }
-        }
-    }
-
-    private func applyLoadedMarkdown(_ text: String, fileURL: URL) {
-        guard currentFileURL?.standardizedFileURL == fileURL.standardizedFileURL else { return }
-        currentMarkdown = text
-        resetAutoSaveFeedback()
-        updateWindowSubtitle()
-        refreshOpenInLLMItem()
-        updateEditToolbarItem()
-        markdownDocument?.replaceContents(markdown: text, fileURL: fileURL)
-        renderCurrentDocument(text: text, fileURL: fileURL)
-        if pendingEditModeURL == fileURL.standardizedFileURL {
-            pendingEditModeURL = nil
-            enterEditMode()
-        }
-    }
-
-    private func applyLoadFailure(error: NSError, fileURL: URL, silentOnFailure: Bool) {
-        if pendingEditModeURL == fileURL.standardizedFileURL {
-            pendingEditModeURL = nil
-            dismissEditChrome()
-        }
-        guard !silentOnFailure else { return }
-        NSAlert(error: error).beginSheetModal(for: documentWindow)
-    }
-
-    private func renderCurrentDocument(text: String, fileURL: URL) {
-        (documentWindow.contentViewController as? MainSplitViewController)?
-            .display(markdown: text,
-                     fileName: fileURL.lastPathComponent,
-                     url: fileURL,
-                     assetBaseURL: fileURL.deletingLastPathComponent())
-    }
-
     /// The formatting bar floats over the editor web view, whose cursor
     /// tracking (I-beam over text, hand over links) otherwise fights the
     /// bar's buttons. The bar region always shows the plain arrow.
-    private final class EditAccessoryContainerView: NSView {
+    final class EditAccessoryContainerView: NSView {
         override func resetCursorRects() {
             addCursorRect(bounds, cursor: .arrow)
         }
@@ -3255,7 +617,7 @@ final class DocumentWindowController: NSWindowController, NSWindowDelegate, NSTo
         }
     }
 
-    private func addBottomTitlebarAccessory(
+    func addBottomTitlebarAccessory(
         _ view: NSView,
         configure: ((NSTitlebarAccessoryViewController) -> Void)? = nil
     ) -> NSTitlebarAccessoryViewController {
@@ -3268,122 +630,4 @@ final class DocumentWindowController: NSWindowController, NSWindowDelegate, NSTo
         return accessory
     }
 
-}
-
-private final class FileWatcher {
-    private static let moveResolutionDelay: TimeInterval = 0.20
-
-    private let url: URL
-    private let onChange: () -> Void
-    /// Fired when the watched file is renamed or moved (in Finder, by an
-    /// editor, etc.). Detected via `F_GETPATH` on the still-open FD —
-    /// the inode follows the file, so the descriptor resolves to the
-    /// new path. Plain deletes don't fire this (path unchanged).
-    var onRename: ((URL) -> Void)?
-    private var source: DispatchSourceFileSystemObject?
-    private var fileDescriptor: Int32 = -1
-    private var debounce: DispatchWorkItem?
-    private var moveResolution: DispatchWorkItem?
-
-    init(url: URL, onChange: @escaping () -> Void) {
-        self.url = url
-        self.onChange = onChange
-        open()
-    }
-
-    private func open() {
-        let fd = Darwin.open(url.path, O_EVTONLY)
-        guard fd >= 0 else { return }
-        fileDescriptor = fd
-
-        let source = DispatchSource.makeFileSystemObjectSource(
-            fileDescriptor: fd,
-            eventMask: [.write, .extend, .delete, .rename, .revoke],
-            queue: .main
-        )
-        source.setEventHandler { [weak self] in
-            guard let self, let source = self.source else { return }
-            let event = source.data
-            // Atomic-rename saves (Vim, VS Code, etc.) replace the inode;
-            // re-open the watcher against the path so we keep tracking.
-            // For an actual user-visible rename, the FD's resolved path
-            // differs from the watcher's URL — surface that to the host.
-            if !event.intersection([.delete, .rename, .revoke]).isEmpty {
-                self.resolveMove(afterSettlingAt: self.currentPath())
-                return
-            }
-            self.scheduleChange()
-        }
-        source.setCancelHandler { [weak self] in
-            guard let self else { return }
-            if self.fileDescriptor >= 0 {
-                Darwin.close(self.fileDescriptor)
-                self.fileDescriptor = -1
-            }
-        }
-        self.source = source
-        source.resume()
-    }
-
-    private func resolveMove(afterSettlingAt movedURL: URL?) {
-        moveResolution?.cancel()
-        debounce?.cancel()
-        source?.cancel()
-        source = nil
-
-        let work = DispatchWorkItem { [weak self] in
-            guard let self else { return }
-            self.moveResolution = nil
-            let resolution = FileWatcherMoveResolution.resolve(
-                originalURL: self.url,
-                movedURL: movedURL,
-                fileExists: { FileManager.default.fileExists(atPath: $0.path) }
-            )
-            switch resolution {
-            case .reloadOriginal:
-                self.open()
-                self.scheduleChange()
-            case .followRename(let newURL):
-                self.onRename?(newURL)
-            case .unavailable:
-                self.reopen()
-            }
-        }
-        moveResolution = work
-        DispatchQueue.main.asyncAfter(
-            deadline: .now() + Self.moveResolutionDelay,
-            execute: work
-        )
-    }
-
-    private func reopen() {
-        source?.cancel()
-        source = nil
-        DispatchQueue.main.asyncAfter(deadline: .now() + 0.05) { [weak self] in
-            self?.open()
-        }
-    }
-
-    private func currentPath() -> URL? {
-        guard fileDescriptor >= 0 else { return nil }
-        var buffer = [CChar](repeating: 0, count: Int(MAXPATHLEN))
-        guard fcntl(fileDescriptor, F_GETPATH, &buffer) == 0 else { return nil }
-        return URL(fileURLWithFileSystemRepresentation: buffer,
-                   isDirectory: false,
-                   relativeTo: nil)
-    }
-
-    private func scheduleChange() {
-        debounce?.cancel()
-        let work = DispatchWorkItem { [weak self] in self?.onChange() }
-        debounce = work
-        DispatchQueue.main.asyncAfter(deadline: .now() + 0.08, execute: work)
-    }
-
-    func cancel() {
-        debounce?.cancel()
-        moveResolution?.cancel()
-        source?.cancel()
-        source = nil
-    }
 }
